% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA to compare three or more experimental groups}
\usage{
ANOVA(
  D,
  id = NULL,
  group,
  sample = NULL,
  paired = FALSE,
  var.equal = TRUE,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  log_base = 2,
  min_obs_per_group = 3,
  min_perc_per_group = NULL,
  verbose = TRUE
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities of the sample.}

\item{id}{\strong{data.frame} \cr
The corresponding ID columns for the parameter D.}

\item{group}{\strong{character factor} \cr
The groups of the data.}

\item{sample}{\strong{character factor} \cr
The sample membership in the data.}

\item{paired}{\strong{logical} \cr
If \code{TRUE}, a paired test will be done, otherwise an unpaired test.}

\item{var.equal}{\strong{logical} \cr
If \code{TRUE}, the variances are assumed to be equal.}

\item{log_before_test}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed.}

\item{delog_for_FC}{\strong{logical} \cr
If \code{TRUE}, the fold change will be calculated without the log-transformation.}

\item{log_base}{\strong{integer} \cr
The base for the log-transformation, if \code{log_before_test = TRUE}.}

\item{min_obs_per_group}{\strong{integer} \cr
The minimum number of observations per group.}

\item{min_perc_per_group}{\strong{integer} \cr
The minimum ratio of observations per group as an alternative to min_obs_per_group.}

\item{verbose}{\strong{logical(1)} \cr
Whether to print messages and a progress bar. Default is TRUE.}
}
\value{
A data.frame with p-values and fold changes
}
\description{
ANOVA to compare three or more experimental groups
}
