% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-normalization.R
\name{automatedNormalization}
\alias{automatedNormalization}
\title{Automated normalization of proteomics data.}
\usage{
automatedNormalization(
  DATA,
  method = "median",
  is_log_transformed = TRUE,
  log_base = 2,
  lts.quantile = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{DATA}{\strong{data.frame} \cr
The data set containing intensities of the sample.}

\item{method}{\strong{character} \cr
The method of normalization. Options are "nonorm" (no normalization), "median", "loess", "quantile" or "lts" normalization.}

\item{is_log_transformed}{\strong{logical} \cr
If \code{TRUE}, the data is log-transformed.}

\item{log_base}{\strong{numeric} \cr
The base, in case the data was log-transformed.}

\item{lts.quantile}{\strong{numeric} \cr
The quantile for the lts normalization.}

\item{verbose}{If TRUE, messages are printed.}
}
\value{
The normalized data as well as a message.
}
\description{
Automated normalization of proteomics data.
}
