% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{calculate_significance_categories}
\alias{calculate_significance_categories}
\title{Calculation of significance categories for a volcano plot}
\usage{
calculate_significance_categories(p, p_adj, fc, thres_fc = 2, thres_p = 0.05)
}
\arguments{
\item{p}{vector of p-values before FDR-correction}

\item{p_adj}{vector of p-values after FDR-correction}

\item{fc}{vector of fold changes}

\item{thres_fc}{threshold for the fold changes}

\item{thres_p}{threshold for the p-values}
}
\value{
a factor with the three significance categories
}
\description{
This function groups all proteins into the following three catrgories based
on the p-values (with and without FDR correction) and fold changes:
\enumerate{
\item not signifcant (p > thres_p or fc < thres_fc)
\item significant (p < thres_p and fc > thres_fc, but p_adj > thres_p)
\item significant after FDR-correction (p_adj < thres_p and fc > thres_fc)
}
}
\examples{
# TODO
}
