% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ttest_ANOVA.R
\name{calculate_significance_categories_ttest}
\alias{calculate_significance_categories_ttest}
\title{Calculation of significance categories for a volcano plot for the t-test.}
\usage{
calculate_significance_categories_ttest(
  p,
  pAdj,
  fc,
  thresFC = 2,
  thresP = 0.05
)
}
\arguments{
\item{p}{\strong{numeric vector} \cr
The p-values before FDR-correction.}

\item{pAdj}{\strong{numeric vector} \cr
The p-values after FDR-correction.}

\item{fc}{\strong{numeric vector} \cr
The values of the fold changes.}

\item{thresFC}{\strong{numeric} \cr
The threshold for the fold changes.}

\item{thresP}{\strong{numeric} \cr
The threshold for the p-values.}
}
\value{
A factor with the three significance categories.
}
\description{
This function groups all proteins into the following three catrgories based
on the p-values (with and without FDR correction) and fold changes:
\enumerate{
\item not signifcant (p > thresP or fc < thresFC)
\item significant (p < thresP and fc > thresFC, but pAdj > thresP)
\item significant after FDR-correction (pAdj < thresP and fc > thresFC)
}
}
