% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_Lineplots.R
\name{clustering}
\alias{clustering}
\title{Cluster proteins for similar patterns across samples}
\usage{
clustering(
  D,
  dist_method = "correlation",
  nr_clusters = NULL,
  cluster_colours = NULL,
  colour_dend = TRUE
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
Dataframe containing protein intensities.}

\item{dist_method}{\strong{character(1)} \cr
Distance measure to use for the hierarchical clustering. In principle,
all methods available in \code{\link[amap]{Dist}} are possible, however
correlation-based metrics like "correlation",  "pearson" or "spearman"
are recommended. The default is "correlation", which uses the centered
Pearson correlation.}

\item{nr_clusters}{\strong{integer(1)} \cr
Number of clusters to cut the dendrogram into. If \code{NULL} (default),
the optimal number of clusters is determined based on silhouette values
using the \code{\link[dendextend]{find_k}} function.}

\item{cluster_colours}{\strong{character vector} \cr
Colours to use for the different clusters. If \code{NULL} (default),
the default ggplot color palette is used.}

\item{colour_dend}{\strong{logical(1)} \cr
If \code{TRUE} (default), the branches of the dendrogram are coloured
according to the clusters, using the defined cluster_colours.}
}
\value{
A list containing the following entries:
\item{row_dend}{The dendrogram object for the rows (proteins).}
\item{nr_clusters}{The number of clusters.}
\item{cluster_colours}{The colours used for the different clusters.}
}
\description{
Cluster proteins for similar patterns across samples
}
