% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_Clustering_Heatmap_Lineplots.R
\name{Clustering_heatmap_lineplots2}
\alias{Clustering_heatmap_lineplots2}
\title{Clustering, Heatmap and Lineplots}
\usage{
Clustering_heatmap_lineplots2(
  D,
  id,
  output_path,
  suffix = "",
  nr_clusters = NULL
)
}
\arguments{
\item{D}{dataframe with log-transformed protein intensities, e.g. filtered for significant proteins form the ANOVA or t-test results}

\item{id}{dataframe with id information e.g. protein names, gene names, accessions etc.}

\item{output_path}{path to save results}

\item{suffix}{suffix for file name}

\item{nr_clusters}{number of clusters. Default is NULL, meaning that the optimal number of clusters will be determined by dendextend::find_k().}
}
\value{
save heatmap and data frame with cluster information, as well as line plots
}
\description{
Clustering, Heatmap and Lineplots
}
\examples{
# TODO
}
