% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ttest_ANOVA.R
\name{.calcSignCat_ttest}
\alias{.calcSignCat_ttest}
\title{Calculation of significance categories for a volcano plot for the t-test.}
\usage{
.calcSignCat_ttest(p, pAdj, fc, thresFC = 2, thresP = 0.05)
}
\arguments{
\item{p}{\strong{numeric}\cr
Vector of p-values before FDR-correction.}

\item{pAdj}{\strong{numeric} \cr
Vector of p-values after FDR-correction.}

\item{fc}{\strong{numeric} \cr
Vector of fold changes.}

\item{thresFC}{\strong{numeric(1)} \cr
Threshold for the fold changes.}

\item{thresP}{\strong{numeric(1)} \cr
Threshold for the p-values.}
}
\value{
A factor vector with the three significance categories.
}
\description{
This function groups all proteins into the following three categories based
on the p-values (with and without FDR correction) and fold changes:
\enumerate{
\item not significant (p > thresP or fc < thresFC)
\item significant (p < thresP and fc > thresFC, but pAdj > thresP)
\item significant after FDR-correction (pAdj < thresP and fc > thresFC)
}
}
\details{
If a p-value is NA, the corresponding significance category will also be NA.
}
