% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MAPlots.R
\name{.MAPlotSingle}
\alias{.MAPlotSingle}
\title{Calculate one single MA plot for comparison of two samples}
\usage{
.MAPlotSingle(sample1, sample2, alpha = 1, pointColour = "black", ...)
}
\arguments{
\item{sample1}{\strong{numeric} \cr
Vector of intensities for the first sample.}

\item{sample2}{\strong{numeric} \cr
Vector of intensities for the second sample..}

\item{alpha}{\strong{numeric(1)} \cr
Transparency factor for data points. Default is 1, no transparency.}

\item{pointColour}{\strong{character} \cr
The color of the data points. Either a single value (colour for all points)
or vector containing a colour for each data point.}

\item{...}{\cr
Additional arguments for \link[affy:ma.plot]{affy::ma.plot}.}
}
\value{
Generates the MA plot for two samples.
}
\description{
Calculate one single MA plot for comparison of two samples
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
D <- prepareData(dataPath = file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)

Intensities <- SummarizedExperiment::assay(D$SE)
s1 <- Intensities[,1]
s2 <- Intensities[,2]

ProtStatsWF:::.MAPlotSingle(sample1 = s1, sample2 = s2)
}
\seealso{
\code{\link[=MAPlots]{MAPlots()}} for calculation of MA-Plots for a whole dataset.
}
