% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{.pivot_longer_SE}
\alias{.pivot_longer_SE}
\title{Pivot SE to long format}
\usage{
.pivot_longer_SE(SE, cols)
}
\arguments{
\item{SE}{A SummarizedExperiment object.}

\item{cols}{Columns to pivot.}
}
\value{
A data frame in long format.
}
\description{
helper function to extract assay data and pivot to long format.
Also adds the information from colData and rowData.
}
