% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-export.R
\name{exportSE}
\alias{exportSE}
\title{Export SummarizedExperiment object to an Excel file, with one assay per sheet.}
\usage{
exportSE(SE, file)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr Object to be exported.}

\item{file}{\strong{character} \cr file path to the output Excel file.
If the file already exists, it will be overwritten.}
}
\value{
nothing, but an Excel file is written containing the assay matrices.
Each assay is written to a separate sheet, and the sheet name corresponds to
the assay name. Row names of the assay matrices are included in the Excel file.
}
\description{
Export SummarizedExperiment object to an Excel file, with one assay per sheet.
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
D_hcc <- prepareData(file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)
exportSE(D_hcc$SE, file = file.path(tempdir(), "HCC_data.xlsx"))
}
