% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PCA.R
\name{filter_PCA_data}
\alias{filter_PCA_data}
\title{A method for filtering the data for PCA.}
\usage{
filter_PCA_data(
  SE,
  assay = "intensity_norm",
  imputeMethod = "none",
  propNA = 0
)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr
The data set containing intensities of the sample.}

\item{assay}{\strong{character(1)} \cr
The name of the assay in SE containing the protein intensities.
Default is "intensity_norm", which uses the normalized
protein intensities if the SE object was generated
by the ProtStatsWF functionality.}

\item{imputeMethod}{\strong{character} \cr
The imputation method. Options are "mean" or "median" or "none}

\item{propNA}{\strong{numeric} \cr
The proportion of allowed missing NAs for a protein, before it is discarded. It is set to 0 if imputeMethod is "none".}
}
\value{
The filtered data.
}
\description{
A method for filtering the data for PCA.
}
