% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Create a Heatmap.}
\usage{
heatmap(
  D,
  id,
  proteinNameColumn = NULL,
  naMethod = "na.omit",
  minValidValues = 2,
  groups = NULL,
  groupColours = NULL,
  columnSplit = NULL,
  clusterColumnSlices = FALSE,
  clusterRows = TRUE,
  clusterColumns = TRUE,
  distMethod = "pearson",
  clustMethod = "complete",
  symmetricLegend = TRUE,
  scaleData = TRUE,
  legendName = "Legend",
  groupName = "Group",
  title = "Heatmap",
  legendColours = c("blue", "white", "red"),
  logData = TRUE,
  logBase = 2,
  colourScaleMax = NULL,
  textSize = 15,
  top_annotation = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities, already filtered for
interesting candidates.}

\item{id}{\strong{data.frame} \cr
The corresponding ID columns for the parameter D e.g. containing further
columns like protein or gene names.}

\item{proteinNameColumn}{\strong{character(1)} \cr
The name of the column in \code{id} with protein or gene names, if the
names should be plotted.}

\item{naMethod}{\strong{character(1)} \cr
The method with which missing values are handled.
Options are "na.omit" (default, proteins with any missing values will be
removed), "impute" (missing values will be imputed) and "keep" (missing
values will be kept). Note that clustering may not work when too many missing
values are present.}

\item{minValidValues}{\strong{integer(1)} \cr
The minimum number of valid values \strong{per row}.
If a protein has less valid values, it will be filtered out.
Note that rows with only 1 or 2 valid values may cause problems with
clustering. Default is 2.}

\item{groups}{\strong{data.frame} \cr
Dataframe containing grouping variables, which are used for column annotations
of the Heatmap. Default is NULL, meaning no column annotations are used.}

\item{groupColours}{\strong{list} \cr
A named list of group colours (discrete vars) or colour functions
(continuous vars), used for the column annotation. Default is NULL, meaning
that random colours will be chosen. See \link[ComplexHeatmap:HeatmapAnnotation]{ComplexHeatmap::HeatmapAnnotation}
for details.}

\item{columnSplit}{\strong{character} \cr  ## ????
The name of the column in the groups dataframe by which the columns should
be split. Default is NULL, then columns will not be splitted.}

\item{clusterColumnSlices}{\strong{logical (1)} \cr
If \code{TRUE}, column slices will be clustered. Default is FALSE.}

\item{clusterRows}{\strong{logical(1) or dendrogram} \cr
If \code{TRUE} (default), the rows will be clustered. Can also be a dendrogram
object which is used to cluster the rows.}

\item{clusterColumns}{\strong{logical or dendrogram} \cr
If \code{TRUE} (default), the columns will be clustered. Can also be a dendrogram
object which is used to cluster the columns.}

\item{distMethod}{\strong{character} \cr
The distance metric for clustering. Options are "pearson", "spearman" and
"euclidean". Default is "pearson".}

\item{clustMethod}{\strong{character} \cr
The linkage method for clustering. Options are "complete", "single" and
"average". Default is "complete".}

\item{symmetricLegend}{\strong{logical} \cr
If \code{TRUE} (default), the colour code will be symmetric.
Note that it only make sense for z-scored data.}

\item{scaleData}{\strong{logical} \cr
If \code{TRUE} (default), the data will be scaled ( = z-scored).}

\item{legendName}{\strong{character(1)} \cr
The name for legend (colour bar). Default is "Legend".}

\item{groupName}{\strong{character} \cr
The name of the group variable, only used for the legend. Default is "Group".}

\item{title}{\strong{character} \cr
The main title of the plot. Default is "Heatmap".}

\item{legendColours}{\strong{character} \cr
A vector of colours for colour gradient.}

\item{logData}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed.}

\item{logBase}{\strong{integer} \cr
The base for the log-transformation, if \code{logData = TRUE}.}

\item{colourScaleMax}{\strong{numeric} \cr
The cap value for which all greater values will receive the same color.}

\item{textSize}{\strong{integer} \cr
The size of text in the plot.}

\item{verbose}{\strong{logical} \cr
If \code{TRUE}, messages will be printed.}

\item{...}{Further arguments to Heatmap}
}
\value{
A heatmap of the given data.
}
\description{
Create a Heatmap.
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
D_hcc <- prepareData(file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)
heatmap(
  D = as.data.frame(SummarizedExperiment::assay(D_hcc$SE, "intensity_norm"))[1:20, ],
  id = as.data.frame(SummarizedExperiment::rowData(D_hcc$SE))[1:20, ],
  proteinNameColumn = "Protein", logData = FALSE, verbose = FALSE
)
}
