% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MAPlots.R
\name{MA_Plot_single}
\alias{MA_Plot_single}
\title{Calculate one MA plot for proteomics data.}
\usage{
MA_Plot_single(
  sample_1,
  sample_2,
  alpha = FALSE,
  pointColour = "black",
  sampling = 1,
  ...
)
}
\arguments{
\item{sample_1}{\strong{numeric vector} \cr
The data of the first sample.}

\item{sample_2}{\strong{numeric vector} \cr
The data of the second sample.}

\item{alpha}{\strong{logical} \cr
If \code{TRUE}, the data points will be transparent.}

\item{sampling}{\strong{numeric} \cr
The sampling rate. Useful to sample part of the data set for data sets on peptide/feature level with many data points.}

\item{...}{Additional arguments for affy::ma.plot.}

\item{point_color}{\strong{character} \cr
The color of the data points.}
}
\value{
Generates the MA plot for two samples.
}
\description{
Calculate one MA plot for proteomics data.
}
\examples{
\dontrun{
prepared_data <- prepareData(...)
data <- prepared_data[["D"]]

s1 <- D[,1]
s1 <- D[,2]

MA_Plot_single(sample_1 = s1, sample_2 = s2)
}

}
\seealso{
\code{\link[=MA_Plots]{MA_Plots()}}
}
