% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MAPlots.R
\name{MA_Plots}
\alias{MA_Plots}
\title{Calculate MA plots for proteomics data set.}
\usage{
MA_Plots(
  D,
  outPath = NULL,
  suffix = "",
  labels = 1:ncol(D),
  labels2 = colnames(D),
  maxPlots = 5000,
  alpha = FALSE,
  plotHeight = 15,
  plotWidth = 15,
  sampling = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing intensities of the sample.}

\item{suffix}{\strong{character} \cr
The suffix, with which the output file will be named. Should start with an underscore.}

\item{labels}{\strong{character} \cr
The sample labels for the title of the MA-Plot.}

\item{labels2}{\strong{character} \cr
The second line in sample title (e.g. group membership).}

\item{maxPlots}{\strong{integer} \cr
The maximum number of MA plots that should be generated.}

\item{alpha}{\strong{logical} \cr
If \code{TRUE}, the data points will be transparent.}

\item{sampling}{\strong{numeric} \cr
The sampling rate. Useful to sample part of the data set for data sets on peptide/feature level with many data points.}

\item{verbose}{If TRUE, messages are printed out.}

\item{...}{Additional arguments for affy::ma.plot.}

\item{output_path}{\strong{character} \cr
The path to a folder for the output.}

\item{plot_height}{\strong{numeric} \cr
The height of the resulting MA plots.}

\item{plot_width}{\strong{numeric} \cr
The width of the resulting MA plots.}
}
\value{
A pdf file containing the MA plots for all sample combinations.
}
\description{
Calculate MA plots for proteomics data set.
}
\examples{
\dontrun{
prepared_data <- prepareData(...)
out_path <- "/Users/thisuser/Documents/resultsFolder/"

return_message <- MA_Plots(D = prepared_data[["D"]], output_path = out_path)
}

}
\seealso{
\code{\link[=MA_Plot_single]{MA_Plot_single()}}
}
