% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_Plot.R
\name{MAPlot_single}
\alias{MAPlot_single}
\title{Calculate MA plots for proteomics data}
\usage{
MAPlot_single(
  x1,
  x2,
  do_log_transformation = TRUE,
  alpha = FALSE,
  col = "black",
  ...
)
}
\arguments{
\item{x1}{}

\item{x2}{}

\item{do_log_transformation}{If \code{TRUE}, the data will be log-transformed.}

\item{alpha}{If \code{TRUE}, the data points will be transparent.}

\item{col}{A character containing the colors of the data points.}

\item{...}{Additional arguments for affy::ma.plot.}
}
\value{
An MA plot for two samples.
}
\description{
Calculate MA plots for proteomics data
}
\examples{
\dontrun{

TODO!!!

}

}
