% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Data preparation}
\usage{
prepareData(
  dataPath,
  intensityColumns,
  proteinNameColumn = "Protein",
  sampleInfoPath = NULL,
  sampleNameColumn = "Sample",
  zeroToNA = TRUE,
  doLogTrans = TRUE,
  logBase = 2,
  normMethod = "loess",
  ltsQuantile = 0.8,
  sep = ",",
  dec = ".",
  header = TRUE,
  sheet = 1,
  NAStrings = c("NA", "NaN", "Filtered", "#NV", ""),
  verbose = TRUE
)
}
\arguments{
\item{dataPath}{\strong{character(1)} \cr
Path to the data file (xlsx, csv, tsv, or txt format) containing the
quantitative proteomics data. The file should have at least a column with
protein names and columns with intensity values for each sample.}

\item{intensityColumns}{\strong{integer} \cr
Column numbers in data that contain the intensity values. E.g. 1:12 if
the first 12 columns contain intensity values for the 12 samples.}

\item{proteinNameColumn}{\strong{character(1)} \cr
Column name in data that contains the protein names. Default is "Protein".}

\item{sampleInfoPath}{\strong{character(1)} \cr
Path to the file containing sample information (optional). Default is NULL
(no sample info file).}

\item{sampleNameColumn}{\strong{character(1)} \cr
Column name of the sampleInfo that contains the sample names. Default is
"Sample". Those names must correspond to column names in data.}

\item{zeroToNA}{\strong{logical(1)} \cr
If TRUE (default), zero intensity values are converted to NA.}

\item{doLogTrans}{\strong{logical(1)} \cr
If TRUE (default), intensity values will be log-transformed. As it is highly
recommended to log-transform proteomics intensities, only set to FALSE if the
data are already log-transformed or you have a special case like TMT ratios.}

\item{logBase}{\strong{numeric(1)} \cr
Base for log transformation. Default is 2.}

\item{normMethod}{\strong{character(1)} \cr
Normalization method, default is "loess". Options are "nonorm", "median",
"quantile", "loess" and "lts".}

\item{ltsQuantile}{\strong{numeric(1)} \cr
Quantile to use for the least trimmed squares regression in the "lts"
normalization method. Default is 0.8.}

\item{sep}{\strong{character(1)} \cr
Separator for csv, tsv, or txt files. Default is "," (comma).
Ignored for xlsx files.}

\item{dec}{\strong{character(1)} \cr
Decimal point character for csv, tsv, or txt files. Default is "." (dot).
Ignored for xlsx files.}

\item{header}{\strong{logical(1)} \cr
If TRUE (default), the first row is treated as column names. Ignored
for xlsx files, where the header is always read.}

\item{sheet}{\strong{integer(1)} \cr
Excel sheet number to read in for dataPath. Only used if fileType = "xlsx".}

\item{NAStrings}{\strong{character} \cr
Character vector of strings to interpret as NA when reading the data file.
Default is c("NA", "NaN", "Filtered","#NV", "").}

\item{verbose}{\strong{logical(1)} \cr
If TRUE (default), messages about the data processing steps are printed.}
}
\value{
List with two elements:
\itemize{
\item SE: A SummarizedExperiment object containing the normalized intensity
data in the assay "intensity_norm", the original intensity data in the assay
"intensity", the protein information in rowData (all other columns in data
that are not the proteinName or intensity columns), and the sample
information in colData.
\item D_long: A data frame in long format containing the normalized
intensity values for each protein and sample, along with the corresponding
sample information from colData.
}
}
\description{
Prepare quantitative proteomics data and combine it with sample information
(optional, e.g., clinical data) in a SummarizedExperiment object.
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
prepareDataSE(dataPath = file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)
}
