% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData2}
\alias{prepareData2}
\title{Title}
\usage{
prepareData2(
  D,
  intensity_columns,
  zero_to_NA = TRUE,
  do_log_transformation = TRUE,
  log_base = 2,
  use_groups = FALSE,
  group_colours = NULL,
  normalization = "loess",
  lts_quantile = 0.8,
  remove_int_below = NULL
)
}
\arguments{
\item{remove_int_below}{remove intensities below this value}
}
\description{
Title
}
