% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_FC_histogram.R
\name{pvalue_foldchange_histogram}
\alias{pvalue_foldchange_histogram}
\title{Generates histograms for p-values, adjusted p-values and fold changes.}
\usage{
pvalue_foldchange_histogram(
  RES,
  columnNameP = "p",
  columnNamePadj = "padj",
  columnNameFC = "FC",
  baseSize = 15
)
}
\arguments{
\item{RES}{\strong{data.frame} \cr
The results from a t-test.}

\item{columnNameP}{\strong{characte} \cr
The column name for p-value.}

\item{columnNamePadj}{\strong{characte} \cr
The columns name for adjusted p-value.}

\item{columnNameFC}{\strong{characte} \cr
The column name for fold change.}
}
\value{
A list of 3 ggplot objects with a histogram for p-values, adjusted p-values and fold changes each.
}
\description{
Generates histograms for p-values, adjusted p-values and fold changes.
}
