% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pvalueFCHistogram.R
\name{pvalueFCHistogram}
\alias{pvalueFCHistogram}
\title{Generates histograms for p-values, adjusted p-values and fold changes.}
\usage{
pvalueFCHistogram(
  RES,
  columnP = "p",
  columnPadj = "padj",
  columnFC = "FC",
  baseSize = 15
)
}
\arguments{
\item{RES}{\strong{data.frame} \cr
The results of a t-test analysis, e.g. generated by \code{\link[=ttest]{ttest()}}. Must contain
at least three columns: the p-value, the adjusted p-value and the fold change.}

\item{columnP}{\strong{character(1)} \cr
Name of column that contains the p-values.}

\item{columnPadj}{\strong{character(1)} \cr
Name of column that contains the adjusted p-values.}

\item{columnFC}{\strong{character(1)} \cr
Name of column that contains the fold changes.}

\item{baseSize}{\strong{numeric(1)} \cr
Base size for the plots. Default is 15.}
}
\value{
A list of 3 ggplot objects with a histogram for p-values,
adjusted p-values and fold changes each.
}
\description{
Generates histograms for p-values, adjusted p-values and fold changes.
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")

D_hcc <- prepareData(dataPath = file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)

RES <- ttest(D_hcc$SE, assay = "intensity_norm",
  groupColumn = "Group", sampleColumn = "PatientID", paired = TRUE,
  logBeforeTest = FALSE, verbose = FALSE)

pvalueFCHistogram(RES, columnP = "p", columnPadj = "p.fdr",
  columnFC = grep("^FC_", names(RES), value = TRUE)[1])
}
