% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-batchnorm.R
\name{regression_normalization}
\alias{regression_normalization}
\title{Batch normalization via linear regression}
\usage{
regression_normalization(
  SE,
  BatchColumn = "Batch",
  pre_norm = TRUE,
  pre_norm_method = "loess",
  return_coefs = FALSE
)
}
\description{
Batch normalization via linear regression
}
