% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ttest.R
\name{ttest_single_row_paired}
\alias{ttest_single_row_paired}
\title{Paired T-test for a single row of a data set.}
\usage{
ttest_single_row_paired(
  x,
  group,
  sample,
  logBeforeTest = TRUE,
  delogForFC = TRUE,
  minNrPairs = NULL,
  logBase = 2,
  row = NULL
)
}
\arguments{
\item{x}{\strong{numeric vector} \cr
The abundances of the data.}

\item{group}{\strong{character factor} \cr
The group membership of the data.}

\item{sample}{\strong{character factor} \cr
The sample membership of the data.}

\item{logBeforeTest}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed before the test.}

\item{delogForFC}{\strong{logical} \cr
If \code{TRUE}, the fold change will be calculated on the original scale.}

\item{minNrPairs}{\strong{integer} \cr
The minimum number of complete sample pairs.}

\item{logBase}{\strong{numeric} \cr
The base of the logarithm for the log-transformation.}

\item{row}{\strong{integer} \cr
The row number of the data for the function call.}
}
\value{
A vector with the following components: mean difference between groups, test statistics, p-value, free space fpr corrected p-value, fold changes (both directions), lower and upper limit of confidence interval, number of valid values per group.
}
\description{
Paired T-test for a single row of a data set.
}
\seealso{
\code{\link[=ttest]{ttest()}}, \code{\link[=ttest_single_row]{ttest_single_row()}}
}
