% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-wilcoxon.R
\name{wilcoxon_single_row_paired}
\alias{wilcoxon_single_row_paired}
\title{Paired Wilcoxon-test for a single row of a data set.}
\usage{
wilcoxon_single_row_paired(
  x,
  group,
  sample,
  delogForFC = TRUE,
  minNrPairs = NULL,
  logBase = 2,
  row = NULL
)
}
\arguments{
\item{x}{\strong{numeric vector} \cr
The abundances of the data.}

\item{group}{\strong{character factor} \cr
The group membership of the data.}

\item{sample}{\strong{character factor} \cr
The sample membership of the data.}

\item{delogForFC}{\strong{logical} \cr
If \code{TRUE}, the fold change will be calculated on the original scale.}

\item{minNrPairs}{\strong{integer} \cr
The minimum number of complete sample pairs.}

\item{logBase}{\strong{numeric} \cr
The base of the logarithm for the log-transformation.}

\item{row}{\strong{integer} \cr
The row number of the data for the function call.}
}
\value{
A vector with the following components: mean difference between groups, test statistics, p-value, free space fpr corrected p-value, fold changes (both directions), lower and upper limit of confidence interval, number of valid values per group.
}
\description{
Paired Wilcoxon-test for a single row of a data set.
}
