% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to a volcano plot.}
\usage{
add_labels(
  RES_Volcano,
  label_type = "FDR",
  ind = NULL,
  protein_name_column = "Gene.names",
  protein_names = NULL
)
}
\arguments{
\item{RES_Volcano}{\strong{ggplot} \cr
The result from \code{\link[=VolcanoPlot]{VolcanoPlot()}}, \code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}} or
\code{\link[=VolcanoPlot_ANOVA]{VolcanoPlot_ANOVA()}}, i.e. a ggplot object containing the transformed data
and significance categories.}

\item{label_type}{\strong{character(1)} \cr
The label type. Options are "FDR" (significant after correction), "noFDR"
(significant without correction) or "index" (define indices to label via
\code{ind}).}

\item{ind}{\strong{integer} \cr
The indices to label if \code{label_type} is "index".}

\item{protein_name_column}{\strong{character(1)} \cr
The column name of the protein names in the \code{RES_Volcano} data.}

\item{protein_names}{\strong{character} \cr
The names of the proteins.}
}
\value{
A ggplot object with labels.
}
\description{
Add labels to a volcano plot.
}
\seealso{
\code{\link[=VolcanoPlot]{VolcanoPlot()}}, \code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=VolcanoPlot_ANOVA]{VolcanoPlot_ANOVA()}}
}
