% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to a volcano plot}
\usage{
add_labels(RES_Volcano, label_type = "FDR", protein_name_column = "Gene.names")
}
\arguments{
\item{RES_Volcano}{result from volcanoPlot(): a list containing the data frame with the transformed data and the ggplot object}

\item{label_type}{"FDR" (significant after correction) or "noFDR" (significant without correction)}

\item{protein_name_column}{column name of the protein names in the RES data frame}
}
\value{
ggplot object with labels
}
\description{
Add labels to a volcano plot
}
\examples{
# TODO
}
