% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to a volcano plot}
\usage{
add_labels(
  RES_Volcano,
  label_type = "FDR",
  ind = NULL,
  protein_name_column = "Gene.names",
  protein_names = NULL
)
}
\arguments{
\item{RES_Volcano}{result from volcanoPlot(): a list containing the data frame with the transformed data and the ggplot object}

\item{label_type}{"FDR" (significant after correction) or "noFDR" (significant without correction) or "index" -> define indizes to label}

\item{protein_name_column}{column name of the protein names in the RES data frame}
}
\value{
ggplot object with labels
}
\description{
Add labels to a volcano plot
}
\examples{
# TODO
}
