% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA to compare three or more experimental groups}
\usage{
ANOVA(
  D,
  id = NULL,
  group,
  sample = NULL,
  paired = FALSE,
  var.equal = TRUE,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  log_base = 2,
  min_obs_per_group = 3,
  min_perc_per_group = NULL,
  filename = "results_ANOVA.xlsx"
)
}
\arguments{
\item{D}{data frame containing only protein or peptide intensities}

\item{id}{data frame containing id columns like protein accessions of gene names}

\item{group}{factor vector indicating the group membership}

\item{sample}{factor vector indicating the sample membership}

\item{paired}{logical indicating whether the test should be paired or unpaired (default is FALSE)}

\item{var.equal}{logical indicating whether the variances of the two groups are assumed to be equal (default is FALSE)}

\item{log_before_test}{logical indicating whether the data should be log-transformed before the test (default is TRUE)}

\item{delog_for_FC}{logical indicating whether the fold change should be calculated on the original scale (default is TRUE)}

\item{log_base}{numeric indicating the base of the logarithm (default is 2)}

\item{min_obs_per_group}{integer indicating the minimum number of observations per group (default is 3)}

\item{min_perc_per_group}{numeric indicating the minimum ratio of observations per group as an alternative to min_obs_per_group}

\item{filename}{character string indicating the name of the output file (default is "results_ANOVA.xlsx")}
}
\value{
data frame with p-values and fold changes
}
\description{
ANOVA to compare three or more experimental groups
}
\examples{
# TODO
}
