% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatedNormalization.R
\name{automatedNormalization}
\alias{automatedNormalization}
\title{Automated normalization of proteomics data}
\usage{
automatedNormalization(
  DATA,
  method = "loess",
  id_columns = NULL,
  groupwise = FALSE,
  group = NULL,
  lts.quantile = 0.8
)
}
\arguments{
\item{DATA}{A data.frame containing the data.}

\item{method}{A character containing the method of normalization. The possible methods are no normalization "nonorm" or "median", "loess", "quantile" or "lts" normalization.}

\item{id_columns}{An integer vector containing the columns of the data of the peptide/protein IDs.}

\item{groupwise}{If \code{TRUE}, the data contains groups.}

\item{group}{The groups of the data, if it has any.}

\item{lts.quantile}{A numeric containing the quantile for the lts normalization.}
}
\value{
The normalized data as well as a message.
}
\description{
Automated normalization of proteomics data
}
