% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-Boxplots.R
\name{Boxplots}
\alias{Boxplots}
\title{Boxplots showing the distribution of intensities over all samples}
\usage{
Boxplots(
  D_long,
  method = "boxplot",
  groupColumn = NULL,
  group_colours = NULL,
  base_size = 15,
  lwd = 0.5,
  outlier_size = 1
)
}
\arguments{
\item{D_long}{\strong{data.frame} \cr
The data set given in long format.}

\item{method}{\strong{character} \cr
The method used. Options are "boxplot" and "violinplot".}

\item{groupColumn}{\strong{logical} \cr}

\item{group_colours}{\strong{character vector} \cr
The hex codes for the group colors.}

\item{base_size}{\strong{numeric} \cr
The base size of the font.}

\item{lwd}{\strong{numeric} \cr
The line width of the boxplot.}

\item{outlier_size}{\strong{numeric} \cr
The size of the outliers.}
}
\value{
boxplots and messages
}
\description{
Boxplots showing the distribution of intensities over all samples
}
\examples{
\dontrun{
prepared_data <- prepareData(...)


boxplot <- Boxplots(D_long = prepared_data[["D_long"]])
}


}
