% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplots.R
\name{Boxplots}
\alias{Boxplots}
\title{Barplots showing the percentage of valid values for each sample.}
\usage{
Boxplots(
  D_long,
  do_log_transformation = FALSE,
  log_base = 2,
  method = "boxplot",
  use_groups = NULL,
  groupvar_name = "Group",
  group_colours = NULL,
  base_size = 15,
  lwd = 0.5,
  outlier_size = 1
)
}
\arguments{
\item{D_long}{\strong{data.frame} \cr
The data set given in long format.}

\item{do_log_transformation}{\strong{logical} \cr
If \code{TRUE}, the data is log-transformed}

\item{log_base}{\strong{numeric} \cr
The base used, if data is log-transformed.}

\item{method}{\strong{character} \cr
The method used. Options are "boxplot" and "violinplot".}

\item{use_groups}{\strong{logical} \cr
If \code{TRUE} data will be plotted in groups.}

\item{groupvar_name}{\strong{character} \cr
The name for the group variable.}

\item{group_colours}{\strong{character vector} \cr
The hex codes for the group colors.}

\item{base_size}{\strong{numeric} \cr
The base size of the font.}

\item{lwd}{\strong{numeric} \cr
The line width of the boxplot.}

\item{outlier_size}{\strong{numeric} \cr
The size of the outliers.}
}
\value{
boxplots and messages
}
\description{
Barplots showing the percentage of valid values for each sample.
}
\examples{
\dontrun{
prepared_data <- prepareData(...)


boxplot <- Boxplots(D_long = prepared_data[["D_long"]])
}


}
