% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot.R
\name{Boxplots}
\alias{Boxplots}
\title{Barplots showing the percentage of valid values for each sample}
\usage{
Boxplots(
  D_long,
  do_log_transformation = FALSE,
  log_base = 2,
  method = "boxplot",
  use_groups = NULL,
  groupvar_name = "Group",
  group_colours = NULL,
  base_size = 15
)
}
\arguments{
\item{D_long}{A data.frame of the data set given in long format.}

\item{do_log_transformation}{If \code{TRUE}, the data is log-transformed}

\item{log_base}{A numeric containing the base used, if data is log-transformed.}

\item{method}{A character containing the method used. Possible are "boxplot" and "violinplot".}

\item{use_groups}{If \code{TRUE} data will be plotted in groups.}

\item{groupvar_name}{A character containing the name for the group variable.}

\item{group_colours}{A character vector of hex codes for the group colors.}

\item{base_size}{A numeric containing the base size of the font.}
}
\value{
a tibble and a ggplot of the valid values
}
\description{
Barplots showing the percentage of valid values for each sample
}
\examples{
\dontrun{
prepared_data <- prepareData(data_path = path, intensity_columns = intensity_cols)

boxplot <- Boxplots(D_long = prepared_data[["D_long"]])
}

}
