% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot_candidates.R
\name{Boxplots_candidates}
\alias{Boxplots_candidates}
\title{Boxplots for biomarker candidates including the data point (jitter).}
\usage{
Boxplots_candidates(
  D,
  proteinNames,
  group = NULL,
  groupColours = NULL,
  logData = TRUE,
  logBase = 2,
  plotDevice = "pdf",
  plotHeight = 15,
  plotWidth = 15,
  plotDPI = 200,
  groupVarName = "Group",
  outputPath,
  suffix = NULL,
  plotNALevel = FALSE
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities,
already filtered for interesting candidates. Usually
the intensities should already be log-transformed.}

\item{proteinNames}{\strong{character vector} \cr
The protein names.}

\item{group}{\strong{character factor} \cr
The group membership of the data.}

\item{groupColours}{\strong{character vector} \cr
The hex codes or colours names for the group colors.}

\item{plotDevice}{\strong{character} \cr
The plot device. Options are "pdf" or "png.}

\item{plotHeight}{\strong{numeric} \cr
The height if plot in cm.}

\item{plotWidth}{\strong{numeric} \cr
The width of plot in cm.}

\item{plotDPI}{\strong{numeric} \cr
The plot resolution (only for png).}

\item{groupVarName}{\strong{character} \cr
The name of the group variable (displayed in legend).}

\item{outputPath}{\strong{character} \cr
The path to a folder for the output.}

\item{suffix}{\strong{character} \cr
The suffix for the output file.}

\item{plotNALevel}{\strong{logical} \cr
If \code{TRUE}, data points will be plotted if the group variable is NA.}
}
\value{
Nothing, saves pdf or png files with boxplots to the output folder.
}
\description{
Boxplots for biomarker candidates including the data point (jitter).
}
