% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplots_biomarker_candidates.R
\name{Boxplots_candidates}
\alias{Boxplots_candidates}
\title{Boxplots for biomarker candidates including the data point (jitter).}
\usage{
Boxplots_candidates(
  D,
  protein.names,
  group = NULL,
  group_colours = NULL,
  log_data = TRUE,
  log_base = 2,
  plot_device = "pdf",
  plot_height = 15,
  plot_width = 15,
  plot_dpi = 200,
  groupvar_name = "Group",
  output_path,
  suffix = NULL,
  plot_NA_level = FALSE
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities, already filtered for interesting candidates.}

\item{protein.names}{\strong{character vector} \cr
The protein names.}

\item{group}{\strong{character factor} \cr
The group membership of the data.}

\item{group_colours}{\strong{character vector} \cr
The hex codes or colours names for the group colors.}

\item{log_data}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed.}

\item{log_base}{\strong{numeric} \cr
The base used, if \code{log_data = TRUE}.}

\item{plot_device}{\strong{character} \cr
The plot device. Options are "pdf" or "png.}

\item{plot_height}{\strong{numeric} \cr
The height if plot in cm.}

\item{plot_width}{\strong{numeric} \cr
The width of plot in cm.}

\item{plot_dpi}{\strong{numeric} \cr
The plot resolution (only for png).}

\item{groupvar_name}{\strong{character} \cr
The name of the group variable (displayed in legend).}

\item{output_path}{\strong{character} \cr
The path to a folder for the output.}

\item{suffix}{\strong{character} \cr
The suffix for the output file.}

\item{plot_NA_level}{\strong{logical} \cr
If \code{TRUE}, data points will be plotted if they are NA.}
}
\value{
Nothing, saves pdf or png files with boxplots to the output folder.
}
\description{
Boxplots for biomarker candidates including the data point (jitter).
}
