% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplots_biomarker_candidates.R
\name{Boxplots_candidates}
\alias{Boxplots_candidates}
\title{Boxplots for biomarker candidates including the data point (jitter)}
\usage{
Boxplots_candidates(
  D,
  protein.names,
  group = NULL,
  group_colours = NULL,
  log_data = TRUE,
  log_base = 2,
  plot_device = "pdf",
  plot_height = 15,
  plot_width = 15,
  plot_dpi = 200,
  groupvar_name = "Group",
  output_path,
  suffix = NULL,
  plot_NA_level = FALSE
)
}
\arguments{
\item{D}{data set containing only protein intensities, already filtered for interesting candidates}

\item{protein.names}{character vector containing protein names}

\item{group}{factor containing the groups}

\item{group_colours}{character vector of hex codes or colours names for the group colors}

\item{log_data}{TRUE, if the data should be log-transformed}

\item{log_base}{base for log-transformation}

\item{plot_device}{plot device, "pdf" or "png}

\item{plot_height}{height if plot in cm}

\item{plot_width}{width of plot in cm}

\item{plot_dpi}{plot resolution (only for png)}

\item{groupvar_name}{name of the group variable (displayed in legend)}

\item{output_path}{output path}

\item{suffix}{suffix for the output file}

\item{plot_NA_level}{TRUE, if data points should be plotted if they are NA}
}
\value{
nothing, saves pdf or png files with boxplots
}
\description{
Boxplots for biomarker candidates including the data point (jitter)
}
\examples{
# TODO
}
