% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplotCandidates.R
\name{BoxplotsCandidates}
\alias{BoxplotsCandidates}
\title{Boxplots for biomarker candidates including the data point (jitter).}
\usage{
BoxplotsCandidates(
  SE,
  assay,
  groupColumn,
  proteinNameColumn,
  groupColours = NULL,
  logData = TRUE,
  logBase = 2,
  plotDevice = "pdf",
  plotHeight = 15,
  plotWidth = 15,
  plotDPI = 200,
  outputPath = NULL,
  suffix = NULL,
  plotNALevel = FALSE
)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr
Data in a SummarizedExperiment object, e.g. output SE from \link{prepareData}
function. It should already be filtered for interesting candidates and
intensities should be log-transformed.}

\item{assay}{\strong{character(1)} \cr
The name of the assay in SE containing the protein intensities that should
be used for the PCA plot. Default is "intensity_norm", which uses the
normalized protein intensities if the SE object was generated by
\link{prepareData}.}

\item{groupColumn}{\strong{character(1)} \cr
The name of the column in the colData of SE containing the groups that will
be compared by the t-test. Please make sure that exactly two groups are
present.}

\item{proteinNameColumn}{\strong{character(1)} \cr
The name of the column in the rowData of SE containing the protein names, e.g.
accessions or gene names. Default is \code{"Protein"}. This column will be
used to label the boxplots.}

\item{groupColours}{\strong{character} \cr
Vector of colours for the groups. Default is NULL (default ggplot2 colour
palette will be used).}

\item{plotDevice}{\strong{character(1)} \cr
Device to use for saving plots. Default is "pdf".}

\item{plotHeight}{\strong{numeric(1)} \cr
Plot height in cm. Default is \code{15}.}

\item{plotWidth}{\strong{numeric(1)} \cr
Plot width in cm. Default is \code{15}.}

\item{plotDPI}{\strong{integer(1)} \cr
Plot resolution in DPI. Default is \code{300}.}

\item{outputPath}{\strong{character(1)} \cr
Path to the output folder. The folder must already exist. If NULL (default),
plots are not saved.}

\item{suffix}{\strong{character(1)} \cr
Suffix to add to the output file names. Default is "". Should ideally start
with an underscore "_".}

\item{plotNALevel}{\strong{logical(1)} \cr
If \code{TRUE}, data points will be plotted if the group variable is NA.
Default is FALSE, so data points with missing group variable will be omitted.}
}
\value{
Nothing, saves pdf or png files with boxplots to the output folder.
}
\description{
Boxplots for biomarker candidates including the data point (jitter).
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv", package = "ProtStatsWF")
file_clinical  <- system.file("extdata", "clinical_data.csv", package = "ProtStatsWF")

D_hcc <- prepareData(dataPath = file_proteins, intensityColumns = 6:43,
                    proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
                    sampleNameColumn = "Sample", verbose = FALSE)

ttest(SE = D_hcc$SE, assay = "intensity_norm",
                  groupColumn = "Group", sampleColumn = "PatientID",,
                  logBeforeTest = FALSE, delogForFC = TRUE, logBase = 2,
                  minObs = 3, paired = TRUE)
fc_col <- paste0("FC_", levels(group_hcc)[[1]], "_divided_by_", levels(group_hcc)[[2]])
sig_cats   <- .calcSignCat_ttest(
 p = ttest_res$p, pAdj = ttest_res$p.fdr, fc = ttest_res[[fc_col]]
)
candidates <- which(as.character(sig_cats) == "significant after FDR correction")

Boxplots_candidates(SE = SE[candidates, ],
assay = assayName,
groupColumn = groupColumn,
proteinNameColumn = proteinNameColumn,
outputPath = getwd())
}
