% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-onoff.R
\name{calculate_onoff}
\alias{calculate_onoff}
\title{Calculate on/off proteins}
\usage{
calculate_onoff(
  D,
  id,
  group,
  maxValidValuesOff,
  minValidValuesOn,
  proteinNamesColumn = 1
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities of the sample.}

\item{id}{\strong{data.frame} \cr
The corresponding ID columns for the parameter D.}

\item{group}{\strong{character factor} \cr
The groups of the data.}

\item{maxValidValuesOff}{\strong{integer} \cr
The proteins below the threshold of the maximum number of valid values are considered off proteins.}

\item{minValidValuesOn}{\strong{integer} \cr
The proteins above the threshold of the minimum number of valid values are considered on proteins.}

\item{proteinNamesColumn}{\strong{character} \cr
The column name in the id dataframe containing the protein IDs used for mapping.}
}
\value{
A data.frame with the number of valid values per group (absolute and relative) and on/off status
}
\description{
Calculate on/off proteins
}
