% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_OnOff_Heatmap.R
\name{calculate_onoff}
\alias{calculate_onoff}
\title{Calculate on/off proteins}
\usage{
calculate_onoff(D, id, group, max_vv_off, min_vv_on, protein_id_col = 1)
}
\arguments{
\item{D}{data set containg only protein intensities}

\item{id}{data frame containing ID columns}

\item{group}{factor containing the groups}

\item{max_vv_off}{off: < max_vv_off valid values}

\item{min_vv_on}{on: > min_vv_on valid values}

\item{protein_id_col}{column on id containing the protein IDs used for mapping}
}
\value{
data frame with number of valid values per group (absolute and relative) and on/off status
}
\description{
Calculate on/off proteins
}
\examples{
# TODO
}
