% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{calculate_significance_categories_ANOVA}
\alias{calculate_significance_categories_ANOVA}
\title{Calculate significance categories for ANOVA}
\usage{
calculate_significance_categories_ANOVA(
  p_posthoc,
  p_anova_adj,
  p_anova,
  fc,
  thres_fc = 2,
  thres_p = 0.05
)
}
\arguments{
\item{fc}{vector of fold changes}

\item{thres_fc}{threshold for the fold changes}

\item{thres_p}{threshold for the p-values}
}
\description{
Calculate significance categories for ANOVA
}
