% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-normalization.R
\name{.normalization}
\alias{.normalization}
\title{Automated normalization of proteomics data.}
\usage{
.normalization(
  DATA,
  method = "loess",
  is_log_transformed = TRUE,
  log_base = 2,
  lts.quantile = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{DATA}{\strong{data.frame} \cr
Data set containing protein or peptide intensities of the samples and no
additional columns.}

\item{method}{\strong{character(1)} \cr
The method of normalization. Options are "nonorm" (no normalization),
"median", "loess" (default), "quantile" or "lts" normalization.}

\item{is_log_transformed}{\strong{logical(1)} \cr
Set to \code{TRUE}, if the data is already log-transformed. Only relevant for lts
normalization.}

\item{log_base}{\strong{numeric(1)} \cr
The base, in case the data was log-transformed before normalization. Default
is 2.}

\item{lts.quantile}{\strong{numeric(1)} \cr
The quantile for the lts normalization. For details check \link[vsn:vsn2]{vsn::vsn2}.}

\item{verbose}{\strong{logical(1)} \cr
If \code{TRUE} (default), messages are printed out.}
}
\value{
The normalized data as a data.frame.
}
\description{
Automated normalization of proteomics data.
}
