% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-PCA.R
\name{.prepDataForPCA}
\alias{.prepDataForPCA}
\title{Prepare data for PCA.}
\usage{
.prepDataForPCA(
  SE,
  assay = "intensity_norm",
  imputeMethod = "none",
  propNA = 0,
  verbose = TRUE
)
}
\arguments{
\item{assay}{\strong{character(1)} \cr
The name of the assay in SE containing the protein intensities that should
be used for the PCA plot. Default is "intensity_norm", which uses the
normalized protein intensities if the SE object was generated by
\link{prepareDataSE}.}

\item{imputeMethod}{\strong{character(1)} \cr
The missing value imputation method. Options are "mean" or "median" or
"none" (default).}

\item{propNA}{\strong{numeric(1)} \cr
The proportion of allowed missing values for a protein, before it is
discarded. It is automatically set to 0 if imputeMethod is "none".}

\item{verbose}{\strong{logical(1)} \cr
If TRUE, messages are printed out.}

\item{**SummarizedExperiment}{object** \cr
Data in a SummarizedExperiment object, e.g. output SE from \link{prepareDataSE}
function.}
}
\value{
SummarizedExperiment object with the filtered data.
}
\description{
Rows (proteins) with too many missing values are removed. If missing values
remain (propNA < 0), the remaining missing values need to be imputed, as the
PCA cannot deal with incomplete data.
Additionally, proteins with almost constant values (and therefor a variance
near zero are removed).
}
