% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ttest.R
\name{.ttest_single_row}
\alias{.ttest_single_row}
\title{Unpaired t-test for a single row of a data set.}
\usage{
.ttest_single_row(
  x,
  group,
  varEqual = FALSE,
  logBeforeTest = FALSE,
  delogForFC = TRUE,
  logBase = 2,
  minObs = 3,
  minObsRatio = NULL,
  row = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{\strong{numeric} \cr
Vector of peptide or protein intensities.}

\item{group}{\strong{factor} \cr
Vector containing the group membership of the samples. Must contain exactly
two factor levels and the same length as \code{x}.}

\item{varEqual}{\strong{logical(1)} \cr
If TRUE, equal variances are assumed (Student's t-test). Default is FALSE
(Welch's t-test).}

\item{logBeforeTest}{\strong{logical(1)} \cr
If TRUE, the data will be log-transformed before the t-test. Default is FALSE.}

\item{delogForFC}{\strong{logical(1)} \cr
If TRUE, the fold change will be calculated on the original scale.
Default is TRUE.}

\item{logBase}{\strong{numeric(1)} \cr
The base of the logarithm for the log-transformation. Default is 2.}

\item{row}{\strong{integer(1)} \cr
The current row number of the data, used for informative warnings.
Default is NULL.}

\item{verbose}{\strong{logical(1)} \cr
If TRUE (default), messages are printed out.}

\item{minObsPerGroup}{\strong{integer(1)} \cr
The minimum number of valid values per group to calculate the test for this
peptide or protein. If fewer valid values are present, NA is reported as
a result. Default is 3.}

\item{minObsPerGroupRatio}{\strong{numeric(1)} \cr
The minimum proportion of valid values required per group (e.g. 0.8 = 80\%
valid values in each group required). This is an alternative to
\code{minObsPerGroup}, especially when groups sizes are unbalanced. If you
use this setting, please set \code{minObsPerGroup = NULL}.}
}
\value{
A named numeric vector with the following elements: mean of group 1,
mean of group 2, test statistic, p-value, placeholder for FDR-corrected
p-value, fold changes in both directions, confidence interval bounds, and
number of valid values per group.
}
\description{
Runs an unpaired t-test on a single row of peptide or protein intensity data
and returns p-values and fold changes.
}
\seealso{
\code{\link[=ttest]{ttest()}}, \code{\link[=.ttest_single_row_paired]{.ttest_single_row_paired()}}
}
