% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_helper.R
\name{filter_PCA_data}
\alias{filter_PCA_data}
\title{A method for filtering the data for PCA.}
\usage{
filter_PCA_data(D, impute = FALSE, impute_method = "mean", propNA = 0)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing intensities of the sample.}

\item{impute}{\strong{logical} \cr
If \code{TRUE}, missing values will be imputed.}

\item{impute_method}{\strong{character} \cr
The imputation method. Options are "mean" or "median".}

\item{propNA}{\strong{numeric} \cr
The proportion of allowed missing NAs for a protein, before it is discarded.}
}
\value{
The filtered data.
}
\description{
A method for filtering the data for PCA.
}
