% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_Clustering_Heatmap_Lineplots.R
\name{getClusterInfos}
\alias{getClusterInfos}
\title{Get cluster information from heatmap}
\usage{
getClusterInfos(heatmap, nr_clusters, D, id)
}
\arguments{
\item{heatmap}{\strong{Heatmap object} \cr
Heatmap object generated by \code{\link[ComplexHeatmap]{Heatmap}}.}

\item{nr_clusters}{\strong{integer(1)} \cr
Number of clusters to cut the dendrogram into.}

\item{D}{\strong{data.frame} \cr
Dataframe containing only protein intensities.}

\item{id}{\strong{data.frame} \cr
Dataframe containing the ID columns for the parameter D e.g. containing further columns like protein or gene names.}
}
\value{
A data.frame containing the cluster assignment for each protein along with the original ID columns and the intensity values.
}
\description{
Get cluster information from heatmap
}
