% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{Heatmap_with_groups}
\alias{Heatmap_with_groups}
\title{Create a Heatmap.}
\usage{
Heatmap_with_groups(
  D,
  id,
  proteinNamesCol = NULL,
  naMethod = "na.omit",
  minValidValues = 2,
  groups = NULL,
  groupColours = NULL,
  columnSplit = NULL,
  clusterColumnSlices = FALSE,
  clusterRows = TRUE,
  clusterColumns = TRUE,
  distMethod = "pearson",
  clustMethod = "complete",
  symmetricLegend = TRUE,
  scaleData = TRUE,
  legendName = "Legend",
  groupName = "Group",
  title = "Heatmap",
  legendColours = c("blue", "white", "red"),
  logData = TRUE,
  logBase = 2,
  colourScaleMax = NULL,
  textSize = 15,
  top_annotation = NULL,
  ...
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities, already filtered for interesting candidates.}

\item{id}{\strong{data.frame} \cr
The corresponding ID columns for the parameter D e.g. containing further columns like protein or gene names}

\item{proteinNamesCol}{\strong{integer} \cr
The column with protein or gene names, if the names should be plotted.}

\item{naMethod}{\strong{character} \cr
The method with which missing values are handeled.
Options are "na.omit" (proteins with any missing values will be removed), "impute" (missing values will be imputed) and "keep" (missing values will be kept).
Note that clustering may not work when too many missing values are present.}

\item{minValidValues}{\strong{integer} \cr
The minimum number of valid values \string{per row}.
If a protein has less valid values, it will be filtered out.
Note that rows with only 1 or 2 valid values may cause problems with clustering.}

\item{groups}{\strong{character factor} \cr
The group of the data with two or more grouping variables.}

\item{groupColours}{\strong{character} \cr
? A named list of group colours (discrete vars) or colour functions (continuous vars)}

\item{columnSplit}{\strong{character} \cr
? The name of the column in groups, if the columns should be split.}

\item{clusterColumnSlices}{\strong{logical} \cr
? If \code{TRUE}, column slices will be clustered.}

\item{clusterRows}{\strong{logical or dendrogram} \cr
If \code{TRUE}, the rows will be clustered. Can also be a dendrogram object which is used to cluster the rows.}

\item{clusterColumns}{\strong{logical or dendrogram} \cr
If \code{TRUE}, the columns will be clustered. Can also be a dendrogram object which is used to cluster the columns.}

\item{distMethod}{\strong{character} \cr
The distance metric for clustering. Options are "pearson", "spearman" and "euclidean".}

\item{clustMethod}{\strong{character} \cr
The linkage method for clustering. Options are "complete", "single" and "average".}

\item{symmetricLegend}{\strong{logical} \cr
If \code{TRUE}, the colour code will be  symmetric.
Note that it only make sense for z-scored data.}

\item{scaleData}{\strong{logical} \cr
If \code{TRUE}, the data will be scaled ( = z-scored).}

\item{legendName}{\strong{character} \cr
The name for legend (colour bar).}

\item{groupName}{\strong{character} \cr
The name of the group variable, only used for the legend.}

\item{title}{\strong{character} \cr
The title of the plot.}

\item{legendColours}{\strong{character} \cr
A vector of colours for colour gradient.}

\item{logData}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed.}

\item{logBase}{\strong{integer} \cr
The base for the log-transformation, if \code{logData = TRUE}.}

\item{colourScaleMax}{\strong{numeric} \cr
The cap value for which all greater values will receive the same color.}

\item{textSize}{\strong{integer} \cr
The size of text in the plot.}

\item{...}{Further arguments to Heatmap}
}
\value{
A heatmap of the given data.
}
\description{
Create a Heatmap.
}
