% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_Heatmap.R
\name{Heatmap_with_groups}
\alias{Heatmap_with_groups}
\title{Heatmap}
\usage{
Heatmap_with_groups(
  D,
  id,
  protein_names_col = NULL,
  na_method = "na.omit",
  filtermissings = 2,
  groups = NULL,
  group_colours = NULL,
  column_split = NULL,
  cluster_column_slices = FALSE,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  dist_method = "pearson",
  clust_method = "complete",
  symmetric_legend = TRUE,
  scale_data = TRUE,
  output_path = paste0(getwd(), "//"),
  suffix = NULL,
  legend_name = "Legend",
  title = "Heatmap",
  legend_colours = c("blue", "white", "red"),
  plot_height = 20,
  plot_width = 20,
  plot_dpi = 300,
  log_data = TRUE,
  log_base = 2,
  colour_scale_max = NULL,
  textsize = 15,
  ...
)
}
\arguments{
\item{D}{dataframe containing only protein intensities (if applicable, already filtered for interesting candidate proteins)}

\item{id}{dataframe containing further columns, e.g. protein or gene names}

\item{protein_names_col}{column with protein or gene names (NULL if no protein names should be plotted)}

\item{na_method}{"na.omit" -> proteins with any missing values will be removed
Please note that clustering may not work when too many missing values are present.}

\item{filtermissings}{filter out proteins with more than X missing values (rows with only 1 or 2 valid values may cause problems with clustering)}

\item{groups}{factor of group or data.frame with two or more grouping variables}

\item{group_colours}{named list of group colours (discrete vars) or colour functions (continuous vars)}

\item{column_split}{should columns be split? NULL if not, name of column in groups if yes}

\item{cluster_column_slices}{cluster the column slices?}

\item{cluster_rows}{logical determining if rows will be clustered. Alternatively a dendrogram can be provided (potentially coloured using the dendextend package).}

\item{cluster_columns}{logical determining if columns will be clustered. Alternatively a dendrogram can be provided (potentially coloured using the dendextend package).}

\item{dist_method}{distance metric for clustering, e.g. "pearson", "spearman", "euclidean"}

\item{clust_method}{linkage method for clustering, e.g. "complete", "single", "average"}

\item{symmetric_legend}{should colour code be made symmetric? (only makes sense for z-scored data)}

\item{scale_data}{should intensities be scaled ( = z-scored per row/protein)? Usually recommended for better visualization of the heatmap.}

\item{output_path}{path for exporting the plot}

\item{suffix}{suffix for file name}

\item{legend_name}{name for legend}

\item{title}{title of the heatmap plot}

\item{legend_colours}{colours for colour gradient (vector of 3 color names of hex codes for low-middle-high intensities)}

\item{plot_height}{plot height}

\item{plot_width}{plot width}

\item{plot_dpi}{plot resolution in DPI}

\item{log_data}{if TRUE, intensities will be log-transformed}

\item{log_base}{base for log-transformation}

\item{colour_scale_max}{maximum value for colour scale (useful if data contains strong outliers)}

\item{textsize}{test size for labels}

\item{...}{further arguments to Heatmap}
}
\value{
heatmap + data used for the heatmap before and after imputation
}
\description{
This function produces a heatmap based on the Heatmap()-function from the ComplexHeatmap package.
Use cases may be to get an overview over the whole proteomics data set or parts of it
(e.g. interesting candidate proteins that were significant in prior analysis).
}
\details{
Missing values

Clustering, Optionen an Methoden (all based on hclust function)

Einfärben weiterer Variablen

weitere Einstelloptionen
}
\examples{
# TODO
}
