% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_Heatmap.R
\name{Heatmap_with_groups}
\alias{Heatmap_with_groups}
\title{Heatmap}
\usage{
Heatmap_with_groups(
  D,
  id,
  protein_names_col = NULL,
  na_method = "na.omit",
  filtermissings = 2,
  groups = NULL,
  group_colours = NULL,
  column_split = NULL,
  cluster_column_slices = FALSE,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  dist_method = "pearson",
  clust_method = "complete",
  symmetric_legend = TRUE,
  scale_data = TRUE,
  output_path = paste0(getwd(), "//"),
  suffix = NULL,
  legend_name = "Legend",
  title = "Heatmap",
  legend_colours = c("blue", "white", "red"),
  plot_height = 20,
  plot_width = 20,
  plot_dpi = 300,
  log_data = TRUE,
  log_base = 2,
  colour_scale_max = NULL,
  textsize = 15,
  ...
)
}
\arguments{
\item{D}{data set containing only protein intensities, already filtered for interesting candidates}

\item{id}{data set containing further columns, e.g. protein or gene names}

\item{protein_names_col}{column with protein or gene names (NULL if no protein names should be plotted)}

\item{na_method}{"na.omit" -> proteins with any missing values will be removed
note that clustering may not work when too many missing values are present}

\item{filtermissings}{filter out proteins with more than X missing values (rows with only 1 or 2 valid values may cause problems with clustering)}

\item{groups}{factor of group or data.frame with two or more grouping variables}

\item{group_colours}{named list of group colours (discrete vars) or colour functions (continuous vars)}

\item{column_split}{should columns be split? NULL if not, name of column in groups if yes}

\item{cluster_column_slices}{cluster the column slices?}

\item{cluster_rows}{if TRUE, rows will be clustered}

\item{cluster_columns}{if TRUE, columns will be clustered}

\item{dist_method}{distance metric for clustering, e.g. "pearson", "spearman", "euclidean"}

\item{clust_method}{linkage method for clustering, e.g. "complete", "single", "average"}

\item{symmetric_legend}{should colour code be made symmetric? (only make sense for z-scored data)}

\item{scale_data}{should data be scaled ( = z-scored)?}

\item{output_path}{path for exporting the plot}

\item{suffix}{suffix for file name}

\item{legend_name}{name for legend}

\item{title}{title}

\item{legend_colours}{colours for colour gradient}

\item{plot_height}{the plot height}

\item{plot_width}{the plot width}

\item{plot_dpi}{the "dots per inch" for the image resolution}

\item{log_data}{If \code{TRUE}, the data will be log-transformed.}

\item{log_base}{The base in which the log-transformation is done.}

\item{colour_scale_max}{the cap value for which all greater values will receive the same color}

\item{textsize}{the size of text in the plot}

\item{...}{further arguments to Heatmap}
}
\value{
heatmap
}
\description{
Heatmap
}
\examples{
# TODO
}
