% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_Lineplots.R
\name{Lineplots}
\alias{Lineplots}
\title{Generate Lineplots for each cluster.}
\usage{
Lineplots(D_zscore, cluster_colours)
}
\arguments{
\item{D_zscore}{\strong{data.frame} \cr
Dataframe containing the z-score normalized protein intensities along
with a column "cluster" indicating the cluster assignment for each protein.}

\item{cluster_colours}{\strong{character vector} \cr
Colours to use for the different clusters.}
}
\value{
A list of ggplot2 objects, each containing the lineplot for one cluster.
}
\description{
Generate Lineplots for each cluster.
}
