% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_Plot.R
\name{MA_Plot_single}
\alias{MA_Plot_single}
\title{Calculate one MA plot for proteomics data}
\usage{
MA_Plot_single(
  sample_1,
  sample_2,
  do_log_transformation = FALSE,
  alpha = FALSE,
  point_color = "black",
  sampling = 1,
  ...
)
}
\arguments{
\item{sample_1}{A numeric vector containing data of the first sample.}

\item{sample_2}{A numeric vector containing data of the second sample.}

\item{do_log_transformation}{If \code{TRUE}, the data will be log-transformed.}

\item{alpha}{If \code{TRUE}, the data points will be transparent.}

\item{point_color}{A character containing the colors of the data points.}

\item{sampling}{A numeric containing the sampling rate. Useful to sample part of the data set for data sets on peptide/feature level with many data points.}

\item{...}{Additional arguments for affy::ma.plot.}
}
\value{
Generates the MA plot for two samples.
}
\description{
Calculate one MA plot for proteomics data
}
\examples{
\dontrun{
prepared_data <- prepareData(...)
data <- prepared_data[["D"]]

s1 <- D[,1]
s1 <- D[,2]

MA_Plot_single(sample_1 = s1, sample_2 = s2)
}

}
