% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_Plot.R
\name{MA_Plots}
\alias{MA_Plots}
\title{Calculate MA plots for proteomics data set}
\usage{
MA_Plots(
  D,
  do_log_transformation = FALSE,
  output_path = "",
  suffix = "",
  labels = 1:ncol(D),
  labels2 = colnames(D),
  maxPlots = 5000,
  alpha = FALSE,
  plot_height = 15,
  plot_width = 15,
  sampling = 1,
  ...
)
}
\arguments{
\item{D}{A data.frame of the data set.}

\item{do_log_transformation}{If \code{TRUE}, the data will be log-transformed.}

\item{output_path}{A character containing the path to a folder.}

\item{suffix}{A character containing the suffix, with which the output file will be named. Should start with an underscore.}

\item{labels}{The sample labels for the title of the MA-Plot.}

\item{labels2}{The second line in sample title (e.g. group membership).}

\item{maxPlots}{A numeric containing the maximum number of MA plots that should be generated.}

\item{alpha}{If \code{TRUE}, the data points will be transparent.}

\item{plot_height}{The height of the resulting MA plots.}

\item{plot_width}{The width of the resulting MA plots.}

\item{sampling}{A numeric containing the sampling rate. Useful to sample part of the data set for data sets on peptide/feature level with many data points.}

\item{...}{Additional arguments for affy::ma.plot.}
}
\value{
A pdf file containing the MA plots for all sample combinations.
}
\description{
Calculate MA plots for proteomics data set
}
\examples{
\dontrun{
prepared_data <- prepareData(...)
out_path <- "/Users/thisuser/Documents/resultsFolder/"

return_message <- MA_Plots(D = prepared_data[["D"]], output_path = out_path)
}

}
