% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MAPlots.R
\name{MAPlots}
\alias{MAPlots}
\title{Calculate MA plots for proteomics data set.}
\usage{
MAPlots(
  D,
  outPath,
  suffix = "",
  labels = as.character(1:ncol(D)),
  labels2 = colnames(D),
  maxPlots = 5000,
  alpha = 1,
  plotHeight = 15,
  plotWidth = 15,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{D}{\strong{matrix} \cr
Data set containing peptide or protein intensities in wide format.}

\item{suffix}{\strong{character(1)} \cr
Suffix for the output file. Should start with an underscore. Default is "".}

\item{labels}{\strong{character} \cr
Vector containing sample labels for each sample. Default is
'as.character(1:ncol(D))', i.e. the respective column number.}

\item{labels2}{\strong{character} \cr
Vector containing a second set of sample labels which are printed as a
second line. Default is 'colnames(D)', i.e. the respective column names.}

\item{maxPlots}{\strong{integer(1)} \cr
The maximum number of MA plots that should be generated. Default is 5000.
This setting reduces the time and amount of space needed for MA-Plots if
the number of samples is high.}

\item{alpha}{\strong{numeric(1)} \cr
Transparency factor for data points. Default is 1, no transparency.}

\item{verbose}{\strong{logical(1)} \cr
If TRUE (default), messages are printed out and a progress bar is shown.}

\item{...}{\cr
Additional arguments for affy::ma.plot.}

\item{outputPath}{\strong{character(1)} \cr
Path to a folder for saving a pdf file with the MA-Plots. Folder must exist.}

\item{plot_height}{\strong{numeric(1)} \cr
The height of the resulting MA plots in cm. Default is 15.}

\item{plot_width}{\strong{numeric(1)} \cr
The width of the resulting MA plots in cm. Default is 15.}
}
\value{
A pdf file containing the MA plots for all sample combinations.
}
\description{
Calculate MA plots for proteomics data set.
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
D <- prepareDataSE(dataPath = file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)

Intensities <- SummarizedExperiment::assay(D$SE)

MAPlots(Intensities)
}
\seealso{
\code{\link[=MAPlotSingle]{MAPlotSingle()}} for internal function that generates a single plot.
}
