% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-PCA.R
\name{PCA_Plot}
\alias{PCA_Plot}
\title{PCA plot for proteomics data.}
\usage{
PCA_Plot(
  SE,
  assay = "intensity_norm",
  groupForColour = NULL,
  colourType = "discrete",
  groupForShape = NULL,
  imputeMethod = "mean",
  propNA = 0,
  scale = TRUE,
  PCx = 1,
  PCy = 2,
  groupColours = NULL,
  groupShapes = NULL,
  NAValueColour = "grey",
  NAValueShape = 0,
  alpha = 1,
  label = FALSE,
  labelSeed = NA,
  labelSize = 4,
  pointSize = 4,
  baseSize = 11,
  xlim = NULL,
  ylim = NULL,
  verbose = TRUE
)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr
Data in a SummarizedExperiment object, e.g. output SE from \link{prepareData}
function.}

\item{assay}{\strong{character(1)} \cr
The name of the assay in SE containing the protein intensities that should
be used for the PCA plot. Default is "intensity_norm", which uses the
normalized protein intensities if the SE object was generated by
\link{prepareData}.}

\item{groupForColour}{\strong{character(1)} \cr
The name of the column in the colData of SE to be used for colouring the
points in the PCA plot. Default is NULL, so no colouring.}

\item{colourType}{\strong{character(1)} \cr
Colour type, either "discrete" (default, discrete set of colours) or
"continuous" (colourbar).}

\item{groupForShape}{\strong{character(1)} \cr
The name of the column in the colData of SE to be used for shaping the points
in the PCA plot.}

\item{imputeMethod}{\strong{character(1)} \cr
The missing value imputation method. Options are "mean" or "median" or
"none" (default).}

\item{propNA}{\strong{numeric(1)} \cr
The proportion of allowed missing values for a protein, before it is
discarded. It is automatically set to 0 if imputeMethod is "none".}

\item{scale}{\strong{logical(1)} \cr
Whether to scale data before calculating the PCA. Default is TRUE.}

\item{PCx}{\strong{integer(1)} \cr
The principal component to be plotted on the x-axis. Default is 1.}

\item{PCy}{\strong{integer(1)} \cr
The principal component to be plotted on the y-axis. Default is 2.}

\item{groupColours}{\strong{character} \cr
Vector of colours for the groups. Default is NULL (default ggplot2 colour
palette will be used).}

\item{groupShapes}{\strong{integer} \cr
Vector of shapes (pch values) for the groups defined by \code{groupForShape}.
Default is NULL (shapes 1, 2, 3, ... are used). Please see
\code{\link[graphics:pch]{graphics::pch()}} for further information.}

\item{NAValueColour}{\strong{character(1)} \cr
Colour for data points with missing values in the groupForColour variable.
Default is "grey".}

\item{NAValueShape}{\strong{integer(1)} \cr
Shape for data points with missing values in the groupForShape variable.
Default is 0 (hollow square). Please see \code{\link[graphics:pch]{graphics::pch()}} for further
information.}

\item{alpha}{\strong{numeric(1)} \cr
Transparency factor for data points. Default is 1, no transparency.}

\item{label}{\strong{logical(1)} \cr
Whether to label the points with the sample names. Default is FALSE.}

\item{labelSeed}{\strong{numeric(1)} \cr
Seed for random number generator used for label placement. Default is NA
(no seed), which may lead to different label placement for multiple
executions of this function.}

\item{labelSize}{\strong{numeric(1)} \cr
Size of the labels. Default is 4.}

\item{pointSize}{\strong{numeric(1)} \cr
Size of data points in the PCA. Default is 4.}

\item{baseSize}{\strong{numeric(1)} \cr
Base size for the plot. Default is 15.}

\item{xlim}{\strong{numeric(2)} \cr
Limits for the x-axis. Default is NULL (automatic).}

\item{ylim}{\strong{numeric(2)} \cr
Limits for the y-axis. Default is NULL (automatic).}

\item{verbose}{\strong{logical(1)} \cr
If TRUE (default), messages are printed out.}
}
\description{
PCA plot for proteomics data.
}
