% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-PCA.R
\name{PCA_Plot}
\alias{PCA_Plot}
\title{PCA plot for proteomics data.}
\usage{
PCA_Plot(
  SE,
  groupForColour = NULL,
  colourType = "discrete",
  groupForShape = NULL,
  assay = "intensity_norm",
  imputeMethod = "mean",
  propNA = 0,
  scale. = TRUE,
  PCx = 1,
  PCy = 2,
  groupColours = NULL,
  alpha = 1,
  label = FALSE,
  labelSeed = NA,
  labelSize = 4,
  xlim = NULL,
  ylim = NULL,
  pointSize = 4,
  baseSize = 11,
  NAValueColour = "grey",
  NAValueShape = 0,
  verbose = TRUE
)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr A SummarizedExperiment object containing the proteomics data (output element SE from \link{prepareDataSE}). The assay specified by the \code{assay} parameter will be used for PCA.}

\item{groupForColour}{\strong{character(1)} \cr The name of the column in the colData of the SummarizedExperiment object to be used for colouring the points in the PCA plot.}

\item{colourType}{\strong{character(1)} \cr Type for colour, either "discrete" (default) or "continuous".}

\item{groupForShape}{\strong{character(1)} \cr The name of the column in the colData of the SummarizedExperiment object to be used for shaping the points in the PCA plot.}

\item{assay}{\strong{character(1)} \cr The name of the assay in the SummarizedExperiment object to be used for PCA.}

\item{imputeMethod}{\strong{character(1)} \cr The method to use for imputing missing values. Options are "mean" (default) and "median".}

\item{propNA}{\strong{numeric} \cr The proportion of missing values allowed before imputation is applied.}

\item{scale.}{\strong{logical(1)} \cr Whether to scale data for the PCA. Default is TRUE.}

\item{PCx}{\strong{integer(1)} \cr The principal component to be plotted on the x-axis. Default is 1.}

\item{PCy}{\strong{integer(1)} \cr The principal component to be plotted on the y-axis. Default is 2.}

\item{groupColours}{\strong{character} \cr Vector of colours for the groups. Default is NULL
(default ggplot2 colour palette will be used.)}

\item{alpha}{\strong{numeric(1)} \cr The transparency level (between 0 and 1) of the points. Default is 1 (no transparency).}

\item{label}{\strong{logical(1)} \cr Whether to label the points. Default is FALSE.}

\item{labelSeed}{\strong{numeric(1)} \cr Seed for random number generator used for label placement.
Default is NA, which may lead to different label placement for multiple executions of this function.}

\item{labelSize}{\strong{numeric(1)} \cr Size of the labels. Default is 4.}

\item{xlim}{\strong{numeric(2)} \cr Limits for the x-axis. Default is NULL (automatic).}

\item{ylim}{\strong{numeric(2)} \cr Limits for the y-axis. Default is NULL (automatic).}

\item{pointSize}{\strong{numeric(1)} \cr Size of data points in the PCA. Default is 4.}

\item{baseSize}{\strong{numeric(1)} \cr Base size for the plots. Default is 15.}

\item{NAValueColour}{\strong{character(1)} \cr Colour for data points with missing values in the groupForColour variable. Default is grey.}

\item{NAValueShape}{\strong{integer(1)} \cr Shape for data points with missing values in the groupForShape variable. Default is 0 (hollow square). Please see \code{\link[graphics:points]{graphics::pch()}} for further information.}

\item{verbose}{\strong{logical(1)} \cr Whether to print messages during the execution of the function. Default is TRUE.}
}
\value{
TODO
}
\description{
PCA plot for proteomics data.
}
