% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Plot.R
\name{PCA_Plot}
\alias{PCA_Plot}
\title{Calculate a PCA plot for proteomics data.}
\usage{
PCA_Plot(
  D,
  groupvar1 = NULL,
  groupvar2 = NULL,
  impute = FALSE,
  impute_method = "mean",
  propNA = 0,
  scale. = TRUE,
  PCx = 1,
  PCy = 2,
  groupvar1_name = "group",
  groupvar2_name = NULL,
  group_colours = NULL,
  alpha = 1,
  label = FALSE,
  label_seed = NA,
  label_size = 4,
  xlim = NULL,
  ylim = NULL,
  point.size = 4,
  base_size = 11,
  ...
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing intensities of the sample.}

\item{groupvar1}{\strong{character vector} \cr
The variable used for colors.}

\item{groupvar2}{\strong{character vector} \cr
The variable used for shapes.}

\item{impute}{\strong{logical} \cr
If \code{TRUE}, missing values will be imputed.}

\item{impute_method}{\strong{character} \cr
The imputation method. Options are "mean" or "median".}

\item{propNA}{\strong{numeric} \cr
The proportion of allowed missing NAs for a protein, before it is discarded.}

\item{scale.}{\strong{logical} \cr
If \code{TRUE}, the data will be scaled before computing the PCA.}

\item{PCx}{\strong{integer} \cr
The principle component for the x-axis.}

\item{PCy}{\strong{integer} \cr
The principle component for the y-axis.}

\item{groupvar1_name}{\strong{character} \cr
The titles of legends for colour.}

\item{groupvar2_name}{\strong{character} \cr
The titles of legends for shape.}

\item{group_colours}{\strong{character vector} \cr
The colors for the groups.}

\item{alpha}{\strong{logical} \cr
If \code{TRUE}, the data points will be transparent.}

\item{label}{\strong{logical} \cr
If \code{TRUE}, the samples will be labeled.}

\item{label_seed}{\strong{numeric} \cr
A numeric, which sets the seed for the label.}

\item{label_size}{\strong{numeric} \cr
A numeric containing the size of the sample labels.}

\item{xlim}{\strong{numeric} \cr
The limit of the x-axis.}

\item{ylim}{\strong{numeric} \cr
The limit of the y-axis.}

\item{point.size}{\strong{numeric} \cr
The size of the data points.}

\item{base_size}{\strong{numeric} \cr
The base size for the plot.}

\item{...}{Additional arguments for ggplot2::plot.}
}
\value{
The PCA plot for a proteomics sample.
}
\description{
Calculate a PCA plot for proteomics data.
}
