% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Plot.R
\name{PCA_Plot}
\alias{PCA_Plot}
\title{Title}
\usage{
PCA_Plot(
  D,
  groupvar1 = NULL,
  groupvar2 = NULL,
  impute = FALSE,
  impute_method = "mean",
  propNA = 0,
  scale. = TRUE,
  PCx = 1,
  PCy = 2,
  groupvar1_name = "group",
  groupvar2_name = NULL,
  group_colours = NULL,
  alpha = 1,
  label = FALSE,
  label_seed = NA,
  label_size = 4,
  xlim = NULL,
  ylim = NULL,
  point.size = 4,
  base_size = 11,
  ...
)
}
\arguments{
\item{D}{A data.frame of the data set.}

\item{groupvar1}{A variable used for colors.}

\item{groupvar2}{A variable used for shapes.}

\item{impute}{If \code{TRUE}, missing values will be imputed.}

\item{impute_method}{A character containing the imputation method ("mean" or "median")}

\item{propNA}{A numeric of the proportion of allowed missing NAs for a protein, before it is discarded.}

\item{scale.}{If \code{TRUE}, the data will be scaled before computing the PCA.}

\item{PCx}{The principle component for the x-axis (default: 1).}

\item{PCy}{The principle component for the y-axis (default: 2).}

\item{groupvar1_name}{Titles of legends for colour and shape.}

\item{groupvar2_name}{Titles of legends for colour and shape.}

\item{group_colours}{The colors for the groups.}

\item{alpha}{If \code{TRUE}, the data points will be transparent.}

\item{label}{If \code{TRUE}, the samples will be labeled.}

\item{label_seed}{A numeric, which sets the seed for the label.}

\item{label_size}{A numeric containing the size of the sample labels.}

\item{xlim}{Limit of the x-axis.}

\item{ylim}{Limit of the y-axis.}

\item{point.size}{The size of the data points.}

\item{base_size}{The base size for the plot.}

\item{...}{Additional arguments for ggplot2::plot.}
}
\value{
The PCA plot for a proteomics sample
}
\description{
Title
}
\examples{
\dontrun{

TODO!!!

}

}
