% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare proteomics data for analysis}
\usage{
prepareData(
  data_path,
  intensity_columns,
  na_strings = c("NA", "NaN", "Filtered", "#NV"),
  zero_to_NA = TRUE,
  log_data = TRUE,
  log_base = 2,
  use_groups = FALSE,
  group_colours = NULL
)
}
\arguments{
\item{data_path}{A character containing the path to an .xlsx file.}

\item{intensity_columns}{An integer vector containing the intensity columns of the table.}

\item{na_strings}{A character vector containing symbols to be recognized as missing values (with the exception of 0).}

\item{log_data}{If \code{TRUE}, the data will be log-transformed.}

\item{log_base}{A numeric containing the base used, if data is log-transformed.}

\item{use_groups}{If \code{TRUE}, the data contains groups.}

\item{group_colours}{A character vector of hex codes for the group colors, if the data has groups. If \code{NULL}, a default color scale will be used.}

\item{zero_to_na}{If \code{TRUE}, 0 will be treated as missing value.}
}
\value{
A list containing the prepared data and the ids of the data as well as the groups, number of groups and group colors
}
\description{
Prepare proteomics data for analysis
}
\examples{
path <- "/User/thisUser/project/data/dataFile.xlsx"
columns <- 3:10
prepareData(data_path = path, intensity_columns = columns)

}
