% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare proteomics data for analysis}
\usage{
prepareData(
  data_path,
  intensity_columns,
  na_strings = c("NA", "NaN", "Filtered", "#NV"),
  zero_to_NA = TRUE,
  do_log_transformation = TRUE,
  log_base = 2,
  use_groups = FALSE,
  group_colours = NULL,
  normalization = "loess"
)
}
\arguments{
\item{data_path}{A character containing the path to an .xlsx file.}

\item{intensity_columns}{An integer vector containing the intensity columns of the table.}

\item{na_strings}{A character vector containing symbols to be recognized as missing values (with the exception of 0).}

\item{zero_to_NA}{If \code{TRUE}, 0 will be treated as missing value.}

\item{do_log_transformation}{If \code{TRUE}, the data will be log-transformed.}

\item{log_base}{A numeric containing the base used, if data is log-transformed.}

\item{use_groups}{If \code{TRUE}, the data contains groups.}

\item{group_colours}{A character vector of hex codes for the group colors, if the data has groups. If \code{NULL}, a default color scale will be used.}

\item{normalization}{A character containing the method of normalization. The possible methods are no normalization "nonorm" or "median", "loess", "quantile" or "lts" normalization.}
}
\value{
A list containing the prepared data and the ids of the data as data.frames as well as the groups, number of groups and group colors
}
\description{
Prepare proteomics data for analysis
}
\examples{
\dontrun{
path <- "/Users/thisuser/Documents/dataFolder/data.xlsx" 
intensity_cols <- 3:17

prepared_data <- prepareData(data_path = path, intensity_columns = intensity_cols)
}

}
