% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare proteomics data for analysis.}
\usage{
prepareData(
  data_path,
  filetype = "xlsx",
  sep = ",",
  dec = ".",
  header = TRUE,
  sheet = 1,
  intensity_columns,
  na_strings = c("NA", "NaN", "Filtered", "#NV"),
  zero_to_NA = TRUE,
  do_log_transformation = TRUE,
  log_base = 2,
  use_groups = FALSE,
  group_colours = NULL,
  normalization = "loess",
  lts_quantile = 0.8
)
}
\arguments{
\item{data_path}{\strong{character} \cr
The path to an .xlsx file containing the input data.}

\item{filetype}{\strong{character(1)} \cr Type of input file: "csv" or "tsv" or "txt" or "xlsx".}

\item{sep}{\strong{character(1)} \cr The field separator, e.g. " " for blanks, "," for comma or "\\t" for tab. Default is ",".}

\item{dec}{\strong{character(1)} \cr Decimal separator, e.g. "," for comma or "." for dot. Default is ".".}

\item{header}{\strong{logical(1)} \cr If TRUE, first line is counted as column names.}

\item{sheet}{\strong{integer(1)} \cr Sheet number (only needed for xlsx files, default is to use the first sheet).}

\item{intensity_columns}{\strong{integer vector} \cr
The numbers of the intensity columns in the table.}

\item{na_strings}{\strong{character} \cr
A vector containing the symbols to be recognized as missing values (with the exception of 0).}

\item{zero_to_NA}{\strong{logical} \cr
If \code{TRUE}, 0 will be treated as missing value.}

\item{do_log_transformation}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed.}

\item{log_base}{\strong{numeric} \cr
The base used, if \code{do_log_transformation = TRUE}.}

\item{use_groups}{\strong{logical} \cr
If \code{TRUE}, group information encoded in the column names is used.}

\item{group_colours}{\strong{character vector} \cr
The hex codes for the group colors, if the data has groups. If \code{NULL}, a default color scale will be used.}

\item{normalization}{\strong{character} \cr
The method of normalization. Options are "nonorm" (no normalization), "median", "loess", "quantile" or "lts" normalization.}

\item{lts_quantile}{\strong{numeric} \cr
The quantile for the lts normalization if \code{normalization = "lts"}.}
}
\value{
A list containing the prepared data and the ids of the data as data.frames as well as the groups, number of groups and group colors.
}
\description{
Prepare proteomics data for analysis.
}
