% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareDataSE}
\alias{prepareDataSE}
\title{Prepare proteomics data for analysis.}
\usage{
prepareDataSE(
  dataPath,
  intensityColumns,
  proteinNameColumn = "Protein",
  sampleInfoPath = NULL,
  sampleNameColumn = "sampleName",
  doLogTrans = TRUE,
  logBase = 2,
  normMethod = "loess",
  ltsQuantile = 0.8,
  fileType = "xlsx",
  sep = ",",
  dec = ".",
  header = TRUE,
  sheet = 1,
  zeroToNA = TRUE,
  NAStrings = c("NA", "NaN", "Filtered", "#NV"),
  verbose = TRUE
)
}
\arguments{
\item{dataPath}{\strong{character(1)} \cr
The path to the data file with protein intensities.}

\item{intensityColumns}{\strong{integer} \cr The columns numbers containing protein intensities in the data set.}

\item{proteinNameColumn}{\strong{character(1)} \cr The name of the column in the data file that contains the protein names. Default is "Protein".}

\item{sampleInfoPath}{\strong{character(1)} \cr The path to the sample information file with group information.
This file must contain a column corresponding to the column names of the data file (only intensity columns).
Further columns are optional and may give information on groups, timepoints or patient information (patient ID, age, gender etc.).}

\item{sampleNameColumn}{\strong{character(1)} \cr The name of the column in the sample information file that contains the sample names.
The sample names have to correspond to the columns names of the data file (only for intensity columns),
but they do not necessarily have to be in the same order as the columns. They will be used to match the data file and the sample information file.}

\item{doLogTrans}{\strong{logical(1)} \cr If \code{TRUE}, the data will be log-transformed.}

\item{logBase}{\strong{numeric(1)} \cr The base for the logarithm, if \code{do_log_transformation = TRUE}. Default is 2.}

\item{normMethod}{\strong{character(1)} \cr The method of normalization. Options are "nonorm" (no normalization), "median", "loess", "quantile" or "lts" normalization.}

\item{ltsQuantile}{\strong{numeric(1)} The quantile for the lts normalization if \code{normalization = "lts"}.}

\item{fileType}{\strong{character(1)} \cr Type of input file: "csv" or "tsv" or "txt" or "xlsx".}

\item{sep}{\strong{character(1)} \cr The field separator, e.g. " " for blanks, "," for comma or "\\t" for tab. Default is ",".}

\item{dec}{\strong{character(1)} \cr Decimal separator, e.g. "," for comma or "." for dot. Default is ".".}

\item{header}{\strong{logical(1)} \cr If TRUE, first line is counted as column names.}

\item{sheet}{\strong{integer(1)} \cr Sheet number (only needed for xlsx files, default is to use the first sheet).}

\item{zeroToNA}{\strong{logical(1)} \cr If \code{TRUE}, 0 will be treated as missing value.}

\item{NAStrings}{\strong{character} \cr A vector containing the symbols to be recognized as missing values (except 0).}

\item{verbose}{\strong{logical(1)} \cr If \code{TRUE}, messages about the performed steps will be printed in the console.}
}
\value{
A list containing the prepared data and the ids of the data as data.frames as well as the groups, number of groups and group colors.
}
\description{
Prepare proteomics data for analysis.
}
