% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ttest_ANOVA.R
\name{prepareTtestData}
\alias{prepareTtestData}
\title{Prepare data from an .xlsx sheet.}
\usage{
prepareTtestData(data_path, intensity_columns)
}
\arguments{
\item{data_path}{\strong{character} \cr
The path to an .xlsx file containing the input data.}

\item{intensity_columns}{\strong{integer vector} \cr
The intensity columns of the table.}
}
\value{
A list containing an intensity data.frame, an IDs data frame and a factor of the sample groups.
}
\description{
Prepare data from an .xlsx sheet.
}
\examples{
\dontrun{
in_path <- "/Users/thisuser/Documents/dataFolder/data.xlsx"
int_cols <- 3:8

result <- workflow_ttest(data_path = in_path, intensity_columns = int_cols)
}
}
