% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_pvalue_FC_histogram.R
\name{pvalue_foldchange_histogram}
\alias{pvalue_foldchange_histogram}
\title{Generates histograms for p-values, adjusted p-values and fold changes.}
\usage{
pvalue_foldchange_histogram(
  RES,
  columnname_p = "p",
  columnname_padj = "padj",
  columnname_FC = "FC",
  base_size = 15
)
}
\arguments{
\item{RES}{\strong{data.frame} \cr
The results from a t-test.}

\item{columnname_p}{\strong{characte} \cr
The column name for p-value.}

\item{columnname_padj}{\strong{characte} \cr
The columns name for adjusted p-value.}

\item{columnname_FC}{\strong{characte} \cr
The column name for fold change.}
}
\value{
A list of 3 ggplot objects with a histogram for p-values, adjusted p-values and fold changes each.
}
\description{
Generates histograms for p-values, adjusted p-values and fold changes.
}
