% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_pvalue_FC_histogram.R
\name{pvalue_foldchange_histogram}
\alias{pvalue_foldchange_histogram}
\title{Generates histograms for p-values, adjusted p-values and fold changes}
\usage{
pvalue_foldchange_histogram(
  RES,
  columnname_p = "p",
  columnname_padj = "padj",
  columnname_FC = "FC"
)
}
\arguments{
\item{RES}{result table from t-test}

\item{columnname_p}{column name for p-value}

\item{columnname_padj}{columns name for adjusted p-value}

\item{columnname_FC}{column name for fold change}
}
\value{
list of 3 ggplot objects
}
\description{
Generates histograms for p-values, adjusted p-values and fold changes
}
\examples{
# TODO
}
