% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_batchnorm.R
\name{regression_normalization}
\alias{regression_normalization}
\title{Regression normalization to remove batch effects}
\usage{
regression_normalization(
  D,
  batch,
  pre_norm = TRUE,
  pre_norm_method = "loess",
  return_coefs = FALSE
)
}
\arguments{
\item{return_coefs}{}
}
\description{
Regression normalization to remove batch effects
}
