% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ttest.R
\name{ttest}
\alias{ttest}
\title{Compute t-test (paired or unpaired)}
\usage{
ttest(
  SE,
  assay,
  groupColumn,
  sampleColumn = NULL,
  paired = FALSE,
  varEqual = FALSE,
  logBeforeTest = TRUE,
  delogForFC = TRUE,
  logBase = 2,
  minObs = 3,
  minObsRatio = NULL,
  verbose = TRUE
)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr
Data in a SummarizedExperiment object, e.g. output SE from \link{prepareData}
function.}

\item{assay}{\strong{character(1)} \cr
The name of the assay in SE containing the protein intensities that should
be used for the PCA plot. Default is "intensity_norm", which uses the
normalized protein intensities if the SE object was generated by
\link{prepareData}.}

\item{groupColumn}{\strong{character(1)} \cr
The name of the column in the colData of SE containing the groups that will
be compared by the t-test. Please make sure that exactly two groups are
present.}

\item{sampleColumn}{\strong{character(1)} \cr
The name of the column in the colData of SE containing the sample IDs. Only
necessary if \code{paired = TRUE}, then it will be used to match sample pairs.}

\item{paired}{\strong{logical(1)} \cr
If TRUE, a paired t-test is performed, otherwise unpaired. Default is FALSE.}

\item{varEqual}{\strong{logical(1)} \cr
If TRUE, equal variances are assumed (Student's t-test). Default is FALSE
(Welch's t-test).}

\item{logBeforeTest}{\strong{logical(1)} \cr
If TRUE, the data will be log-transformed before the t-test. Default is FALSE.}

\item{delogForFC}{\strong{logical(1)} \cr
If TRUE, the fold change will be calculated on the original scale.
Default is TRUE.}

\item{logBase}{\strong{numeric(1)} \cr
The base of the logarithm for the log-transformation. Default is 2.}

\item{minObs}{\strong{integer(1)} \cr
Minimum number of observations per group. For a paired t-test this is the
minimum number of complete pairs. Default is 3.}

\item{minObsRatio}{\strong{numeric(1)} \cr
Minimum proportion of valid values required per group (e.g. 0.8 = 80\%).
Currently not usable if \code{paired = TRUE}.}

\item{verbose}{\strong{logical} \cr
If \code{TRUE}, messages will be printed.}
}
\value{
A data frame containing the t-test results an used protein intensity
values.
}
\description{
Applies an unpaired or paired t-test row-wise across a matrix of protein
intensities and returns a combined data frame of input data and test results
with FDR-adjusted p-values.
}
\examples{
file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
D_hcc <- prepareData(file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)
ttest(D_hcc$SE, assay = "intensity_norm", groupColumn = "Group",
  sampleColumn = "PatientID", paired = TRUE, logBeforeTest = FALSE,
  verbose = FALSE)
}
\seealso{
For more details see \code{\link[=.ttest_single_row]{.ttest_single_row()}} for the unpaired and
\code{\link[=.ttest_single_row_paired]{.ttest_single_row_paired()}} for the paired t-test.
}
