% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ttest.R
\name{ttest}
\alias{ttest}
\title{Compute t-test (paired or unpaired)}
\usage{
ttest(
  SE,
  assay,
  groupColumn,
  sampleColumn = NULL,
  paired = FALSE,
  varEqual = FALSE,
  logBeforeTest = TRUE,
  delogForFC = TRUE,
  logBase = 2,
  minObs = 3,
  minObsRatio = NULL,
  verbose = TRUE
)
}
\arguments{
\item{SE}{\strong{SummarizedExperiment object} \cr
Data in a SummarizedExperiment object, e.g. output SE from \link{prepareDataSE}
function.}

\item{assay}{\strong{character(1)} \cr
The name of the assay in SE containing the protein intensities that should
be used for the PCA plot. Default is "intensity_norm", which uses the
normalized protein intensities if the SE object was generated by
\link{prepareDataSE}.}

\item{groupColumn}{\strong{character(1)} \cr
The name of the column in the colData of SE containing the groups that will
be compared by the t-test. Please make sure that exactly two groups are
present.}

\item{sampleColumn}{\strong{character(1)} \cr
The name of the column in the colData of SE containing the sample IDs. Only
necessary if \code{paired = TRUE}, then it will be used to match sample pairs.}

\item{paired}{\strong{logical(1)} \cr
If TRUE, a paired t-test is performed, otherwise unpaired. Default is FALSE.}

\item{minObs}{\strong{integer(1)} \cr
Minimum number of observations per group. For a paired t-test this is the
minimum number of complete pairs. Default is 3.}

\item{minObsRatio}{\strong{numeric(1)} \cr
Minimum proportion of valid values required per group (e.g. 0.8 = 80\%).
Currently not usable if \code{paired = TRUE}.}
}
\value{
A data frame containing the t-test results an used protein intensity
values.
}
\description{
Applies an unpaired or paired t-test row-wise across a matrix of protein
intensities and returns a combined data frame of input data and test results
with FDR-adjusted p-values.
}
\seealso{
\code{\link[=.ttest_single_row]{.ttest_single_row()}}, \code{\link[=.ttest_single_row_paired]{.ttest_single_row_paired()}}
}
