% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Function to compute t-test (paired or unpaired)}
\usage{
ttest(
  D,
  id = NULL,
  group,
  sample = NULL,
  paired = FALSE,
  var.equal = FALSE,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  log_base = 2,
  min_obs_per_group = 3,
  filename = "results_ttest.xlsx",
  min_obs_per_group_ratio = NULL
)
}
\arguments{
\item{D}{\strong{data.frame} \cr
The data set containing only protein intensities, already filtered for interesting candidates.}

\item{id}{\strong{data.frame} \cr
The corresponding ID columns for the parameter D e.g. containing further columns like protein or gene names.}

\item{group}{\strong{character factor} \cr
The group membership of the data.}

\item{sample}{\strong{character factor} \cr
The sample membership of the data.}

\item{paired}{\strong{logical} \cr
If \code{TRUE}, the test will be paired, otherwise it will be unpaired.}

\item{var.equal}{\strong{logical} \cr
If \code{TRUE}, the variances of the groups are expected to be equal.}

\item{log_before_test}{\strong{logical} \cr
If \code{TRUE}, the data will be log-transformed before the test.}

\item{delog_for_FC}{\strong{logical} \cr
If \code{TRUE}, the fold change will be calculated on the original scale.}

\item{log_base}{\strong{numeric} \cr
The base of the logarithm for the log-transformation.}

\item{min_obs_per_group}{\strong{integer} \cr
The minimum number of observations per group.}

\item{filename}{\strong{characte} \cr
The name of the output file.}

\item{min_obs_per_group_ratio}{\strong{numeric} \cr
The minimum number of observations per group as a ratio (e.g, 0.8 = 80\% valid values in each group needed).}
}
\value{
A data frame containing the results of the t-test.
}
\description{
Function to compute t-test (paired or unpaired)
}
\seealso{
\code{\link[=ttest_single_row]{ttest_single_row()}}, \code{\link[=ttest_single_row_paired]{ttest_single_row_paired()}}
}
