% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Function to compute t-test (paired or unpaired)}
\usage{
ttest(
  D,
  id = NULL,
  group,
  sample = NULL,
  paired = FALSE,
  var.equal = FALSE,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  log_base = 2,
  min_obs_per_group = 3,
  filename = "results_ttest.xlsx",
  min_obs_per_group_ratio = NULL
)
}
\arguments{
\item{D}{data frame containing only protein or peptide intensities}

\item{id}{data frame containing id columns like protein accessions of gene names}

\item{group}{factor vector indicating the group membership}

\item{sample}{factor vector indicating the sample membership}

\item{paired}{logical indicating whether the test should be paired or unpaired (default is FALSE)}

\item{var.equal}{logical indicating whether the variances of the two groups are assumed to be equal (default is FALSE)}

\item{log_before_test}{logical indicating whether the data should be log-transformed before the test (default is TRUE)}

\item{delog_for_FC}{logical indicating whether the fold change should be calculated on the original scale (default is TRUE)}

\item{log_base}{numeric indicating the base of the logarithm (default is 2)}

\item{min_obs_per_group}{integer indicating the minimum number of observations per group (for unpaired test) or minimum number of samples pairs (for paored test) (default is 3)}

\item{filename}{character string indicating the name of the output file (default is "results_ttest.xlsx")}

\item{min_obs_per_group_ratio}{numeric indicating the minimum ratio of observations per group (for unpaired test) as an alternative to min_obs_per_group}
}
\value{
a data frame containing the results of the t-test
}
\description{
Function to compute t-test (paired or unpaired)
}
\examples{
# TODO
}
