% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_single_row}
\alias{ttest_single_row}
\title{Unpaired t-test for a single row of a data set}
\usage{
ttest_single_row(
  x,
  group,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  min_obs_per_group = NULL,
  min_obs_per_group_ratio = NULL,
  log_base = 2,
  row = NULL,
  var.equal = FALSE
)
}
\arguments{
\item{x}{A numeric vector of abundances.}

\item{group}{A factor vector of group membership.}

\item{log_before_test}{A logical indicating whether the data should be log-transformed before the test.}

\item{delog_for_FC}{A logical indicating whether the fold change should be calculated on the original scale.}

\item{min_obs_per_group}{An integer indicating the minimum number of observations per group.}

\item{min_obs_per_group_ratio}{A numeric indicating the minimum number of observations per group as a ratio (e.g, 0.8 = 80\% valid values in each group needed).}

\item{log_base}{A numeric indicating the base of the logarithm.}

\item{row}{A numeric indicating the row number.}

\item{var.equal}{A logical indicating whether the variances of the groups are expected to be equal.}
}
\value{
A vector with the following components: mean group 1, mean group 2, test statistics, p-value, free space fpr corrected p-value, fold changes (both directions), lower and upper limit of confidence interval, number of valid values per group.
}
\description{
Unpaired t-test for a single row of a data set
}
\examples{
# TODO
}
