% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest_single_row_paired}
\alias{ttest_single_row_paired}
\title{Paired T-test for a single row of a data set}
\usage{
ttest_single_row_paired(
  x,
  group,
  sample,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  min_nr_pairs = NULL,
  log_base = 2,
  row = NULL
)
}
\arguments{
\item{x}{A numeric vector of abundances.}

\item{group}{A factor vector of group membership.}

\item{sample}{A vector indicating sample pairs.}

\item{log_before_test}{A logical indicating whether the data should be log-transformed before the test.}

\item{delog_for_FC}{A logical indicating whether the fold change should be calculated on the original scale.}

\item{min_nr_pairs}{An integer indicating the minimum number of complete sample pairs.}

\item{log_base}{A numeric indicating the base of the logarithm.}

\item{row}{A numeric indicating the row number.}
}
\value{
A vector with the following components: mean difference between groups, test statistics, p-value, free space fpr corrected p-value, fold changes (both directions), lower and upper limit of confidence interval, number of valid values per group.
}
\description{
Paired T-test for a single row of a data set
}
\examples{
# TODO
}
