% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-valid_values.R
\name{ValidValuePlot}
\alias{ValidValuePlot}
\title{Barplots showing the percentage of valid values for each sample.}
\usage{
ValidValuePlot(D_long, groupColumn = NULL, groupColours = NULL, baseSize = 15)
}
\arguments{
\item{D_long}{\strong{data.frame} \cr
The data set given in long format.}

\item{groupColumn}{\strong{character(1)} \cr
Column used for colouring.}

\item{group_colours}{\strong{character vector} \cr
The hex codes for the group colors.}

\item{base_size}{\strong{numeric} \cr
The base size of the font.}
}
\value{
A tibble and a ggplot of the valid values.
}
\description{
Barplots showing the percentage of valid values for each sample.
}
\examples{

#path <- system.file("extdata", "exampleQC.xlsx", package = "ProtStatsWF")
#columns <- 3:11

#prepared_data <- prepareData(data_path = path, intensity_columns = columns)

#vvplot <- ValidValuePlot(D_long = prepared_data[["D_long"]])

#plot(vvplot[["plot"]])


}
