% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-valid_values.R
\name{ValidValuePlot}
\alias{ValidValuePlot}
\title{Barplots showing the number of valid values for each sample.}
\usage{
ValidValuePlot(D_long, groupColumn = NULL, groupColours = NULL, baseSize = 15)
}
\arguments{
\item{D_long}{\strong{data.frame} \cr
The data set in long format, output object from \link{prepareData}.
Must at least contain the columns ".sample" (sample IDs) and
"intensity_norm" (protein intensities).
May contain additional columns that can be used for assigning samples to
groups for colouring.}

\item{groupColumn}{\strong{character(1)} \cr
Name of the column in D_long that contains the group information for the
samples. If NULL (default), no grouping is applied.}

\item{groupColours}{\strong{character} \cr
Vector containing names of hex codes for the group colours (assigned in
alphabetical order of the groups). If NULL (default), ggplot2 default
colours are used.}

\item{baseSize}{\strong{numeric(1)} \cr
The base size of the font used for axis labels etc. Default is 15.}
}
\value{
A list with two elements:
\itemize{
\item plot: ggplot2 object of the valid value plot
\item table: table containing the plotted numbers
}
}
\description{
Barplots showing the number of valid values for each sample.
}
\examples{

file_proteins <- system.file("extdata", "proteins_HCC.csv",
  package = "ProtStatsWF")
file_clinical <- system.file("extdata", "clinical_data.csv",
  package = "ProtStatsWF")
D <- prepareDataSE(dataPath = file_proteins, intensityColumns = 6:43,
  proteinNameColumn = "Protein", sampleInfoPath = file_clinical,
  sampleNameColumn = "Sample", verbose = FALSE)

ValidValuePlot(D_long = D$D_long, groupColumn = "Group")

}
