% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidValuePlot.R
\name{ValidValuePlot}
\alias{ValidValuePlot}
\title{Barplots showing the percentage of valid values for each sample}
\usage{
ValidValuePlot(
  D_long,
  output_path = "",
  suffix = "",
  groupvar_name = "Group",
  group_colours = NULL,
  plot_device = "png",
  plot_height = 10,
  plot_width = 15,
  plot_dpi = 300,
  base_size = 15
)
}
\arguments{
\item{D_long}{A data.frame of the data set given in long format.}

\item{output_path}{A character containing the output path .}

\item{suffix}{A character containing the suffix for the file name.}

\item{groupvar_name}{A character containing the name for the group variable.}

\item{group_colours}{A character vector of hex codes for the group colors.}

\item{plot_device}{A character containing the plot device to save graphics.}

\item{plot_height}{A numeric containing the height of the plot.}

\item{plot_width}{A numeric containing the width of the plot.}

\item{plot_dpi}{A numeric containing the resolution of the plot (in dots per inch).}

\item{base_size}{A numeric containing the base size of the font.}
}
\value{
ggplot, saves a table
}
\description{
Barplots showing the percentage of valid values for each sample
}
