% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidValuePlot.R
\name{ValidValuePlot}
\alias{ValidValuePlot}
\title{Barplot for number of valid values}
\usage{
ValidValuePlot(
  D_long,
  output_path = "",
  suffix = "",
  groupvar_name = "Group",
  group_colours = NULL,
  plot_device = "png",
  plot_height = 10,
  plot_width = 15,
  plot_dpi = 300,
  base_size = 15
)
}
\arguments{
\item{D_long}{data set in long format}

\item{output_path}{output path}

\item{suffix}{suffix for file name}

\item{groupvar_name}{name for the group variable}

\item{group_colours}{colours for the groups}

\item{plot_device}{device}

\item{plot_height}{height of the plot}

\item{plot_width}{width of the plot}

\item{plot_dpi}{resolution of the plot}

\item{base_size}{base size of the font size}
}
\value{
ggplot, saves a table
}
\description{
Barplot for number of valid values
}
