% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Simple volcano plot from p-value, fold change and significance category}
\usage{
VolcanoPlot(
  p,
  FC,
  significance_category,
  logBaseFC = 2,
  logBaseP = 10,
  thresP = 0.05,
  thresFC = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetricX = TRUE,
  legendPosition = "bottom",
  baseSize = 15,
  xlim = NULL,
  ylim = NULL,
  alpha = 0.5,
  pointSize = 3,
  linewidth = 2,
  showThresLine = TRUE
)
}
\arguments{
\item{p}{\strong{numeric} \cr
Vector containing p-values.}

\item{FC}{\strong{numeric} \cr
Vector containing fold changes.}

\item{logBaseFC}{\strong{numeric(1)} \cr
The base for the log-transformation of fold changes. Default is 2.}

\item{logBaseP}{\strong{numeric(1)} \cr
The base for the log-transformation of p values. Default is 10}

\item{thresP}{\strong{numeric(1)} \cr
Threshold for the p-values, used to draw a horizontal line. Default is 0.05.}

\item{thresFC}{\strong{numeric()} \cr
Threshold for the fold changes, used to draw two vertical lines at thresFC and
1/thresFC. Default is 2.}

\item{colour1}{\strong{character(1)} \cr
The color for not significant proteins. Default is "grey".}

\item{colour2}{\strong{character(1)} \cr
The color for significant proteins. Defaults is "black".}

\item{colour3}{\strong{character(1)} \cr
The color for significant proteins after FDR correction. Default is "orange".}

\item{symmetricX}{\strong{logical(1)} \cr
If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is
defined). Default is TRUE.}

\item{legendPosition}{\strong{character(1)} \cr
The positioning of the legend. Options are "none", "left", "right", "bottom",
"top" and "inside".}

\item{baseSize}{\strong{numeric(1)} \cr
Base size for the plots. Default is 15.}

\item{xlim}{\strong{numeric(2)} \cr
The limits for x-axis. Default is NULL, i.e. will be determined automatically.}

\item{ylim}{\strong{numeric(2)} \cr
The limits for y-axis. Default is NULL, i.e. will be determined automatically.}

\item{alpha}{\strong{numeric(1)} \cr
The degree of transparency of the data points, between 0 and 1. Default is 1,
i.e. no transparency.}

\item{pointSize}{\strong{numeric(1)} \cr
The size of the data points. Default is 3.}

\item{linewidth}{\strong{numeric(1)} \cr
The width of the threshold lines. Default is 2.}

\item{showThresLine}{\strong{logical(1)} \cr
If \code{TRUE}, threshold lines are drawn at \code{thresP} and at
\code{thresFC} (and its inverse). Default is TRUE.}

\item{significanceCategories}{\strong{factor} \cr
The significance categories for the volcano plot, e.g. generated by
\link{.calcSignCat_ttest}}
}
\value{
A ggplot object containing the volcano plot.
}
\description{
Simple volcano plot from p-value, fold change and significance category
}
\examples{
p <- c(0.001, 0.02, 0.4)
FC <- c(2.5, 0.4, 1.1)
significance <- factor(
  c("significant after FDR correction", "significant", "not significant"),
  levels = c("not significant", "significant",
    "significant after FDR correction")
)
VolcanoPlot(p, FC, significance)
}
\seealso{
\code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=VolcanoPlot_ANOVA]{VolcanoPlot_ANOVA()}}
}
