% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Simple volcano plot from p-value, fold change and significance category}
\usage{
VolcanoPlot(
  p,
  FC,
  significance_category,
  logBaseFC = 2,
  logBaseP = 10,
  thresP = 0.05,
  thresFC = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  alpha = 0.5,
  point_size = 3,
  linewidth = 2
)
}
\arguments{
\item{p}{\strong{numeric factor} \cr
The values for the p-values.}

\item{FC}{\strong{numeric factor} \cr
The values for the fold changes.}

\item{significance_category}{\strong{character factor} \cr
The significance categories for the volcano plot.}

\item{colour1}{\strong{character} \cr
The color for not significant proteins.}

\item{colour2}{\strong{character} \cr
The color for significant proteins.}

\item{colour3}{\strong{character} \cr
The color for significant proteins after FDR correction.}

\item{symmetric_x}{\strong{logical} \cr
If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is defined).}

\item{legend_position}{\strong{character} \cr
The positioning of the legend.
Options are "none", "left", "right", "bottom", "top" and "inside".}

\item{base_size}{\strong{numeric} \cr
The base size of the font.}

\item{xlim}{\strong{numeric} \cr
The limits for x-axis.}

\item{ylim}{\strong{numeric} \cr
The limits for y-axis.}

\item{alpha}{\strong{numeric} \cr
The transparency of the data points.}

\item{point_size}{\strong{numeric} \cr
The size of the data points.}

\item{log_base_fc}{\strong{numeric} \cr
The base for the fold changes log-transformation.}

\item{log_base_p}{\strong{numeric} \cr
The base for the p-values log-transformation.}

\item{thres_p}{\strong{numeric} \cr
The threshold for the p-values.}

\item{thres_fc}{\strong{numeric} \cr
The threshold for the fold changes.}
}
\value{
A ggplot showing the volcano plot.
}
\description{
Simple volcano plot from p-value, fold change and significance category
}
\seealso{
\code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=VolcanoPlot_ANOVA]{VolcanoPlot_ANOVA()}}
}
