% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Simple volcano plot from p-value, fold change and significance category}
\usage{
VolcanoPlot(
  p,
  FC,
  significance_category,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{p}{Vector containing the p-values}

\item{FC}{Vector containing the fold changes}

\item{significance_category}{}

\item{log_base_fc}{Numeric of the base for the fold changes log-transformation}

\item{log_base_p}{Numeric of the base for the p-values log-transformation}

\item{thres_p}{Numeric of the threshold for the p-values}

\item{thres_fc}{Numeric of the threshold for the fold changes}

\item{colour1}{Character of the color for not significant proteins}

\item{colour2}{Character of the color for significant proteins}

\item{colour3}{Character of the color for significant proteins after FDR correction}

\item{symmetric_x}{If TRUE, x-axis limits will be made symmetric (not used if xlim is defined)}

\item{legend_position}{Character of the positioning of the legend, e.g. "bottom" or "right"}

\item{base_size}{Numeric containing the base size of the font}

\item{xlim}{The limits for x-axis. Default is NULL}

\item{ylim}{The limits for y-axis. Default is NULL}
}
\description{
Simple volcano plot from p-value, fold change and significance category
}
