% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano plot}
\usage{
VolcanoPlot(
  RES,
  columnname_p = "p",
  columnname_padj = "padj",
  columnname_FC = "FC",
  log_base_fc = 2,
  log_base_p = 10,
  is_FC_log = FALSE,
  is_p_log = FALSE,
  thres_fc = 2,
  thres_p = 0.05,
  show_thres_line = TRUE,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  groupname1 = "group1",
  groupname2 = "group2",
  xlim = NULL,
  ylim = NULL,
  symmetric_x = FALSE,
  legend_position = "bottom",
  plot_height = 15,
  plot_width = 15,
  plot_dpi = 300,
  plot_device = "pdf",
  output_path = NULL,
  suffix = NULL,
  base_size = NULL,
  add_annotation = TRUE
)
}
\arguments{
\item{RES}{result table from t-test}

\item{columnname_p}{column name for p-value}

\item{columnname_padj}{columns name for adjusted p-value}

\item{columnname_FC}{column name for fold change}

\item{log_base_fc}{base for log transformation of fold change}

\item{log_base_p}{base for log transformation of p-value}

\item{is_FC_log}{TRUE if fold change is already log-transformed}

\item{is_p_log}{FALSE if p-value is already log-transformed}

\item{thres_fc}{threshold for fold change}

\item{thres_p}{threshold for p-value}

\item{show_thres_line}{TRUE if threshold lines should be shown}

\item{colour1}{colour for not significant proteins}

\item{colour2}{colour for significant proteins}

\item{colour3}{colour for significant proteins after FDR correction}

\item{groupname1}{name of first group}

\item{groupname2}{name of second group}

\item{xlim}{limits for x-axis}

\item{ylim}{limits for y-axis}

\item{symmetric_x}{if TRUE, x-axis limits will be made symmetric (not used if xlim is defined)}

\item{legend_position}{position of the legend (e.g., "bottom" or "right")}

\item{plot_height}{height of plot}

\item{plot_width}{width of plot}

\item{plot_dpi}{resolution of plot}

\item{plot_device}{plot device}

\item{output_path}{path for output}

\item{suffix}{suffix for output file}

\item{base_size}{base size for theme}

\item{add_annotation}{TRUE if annotation should be added}
}
\value{
ggplot object
}
\description{
Volcano plot
}
\examples{
# TODO
}
