% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Simple volcano plot from p-value, fold change and significance category}
\usage{
VolcanoPlot(
  p,
  FC,
  significance_category,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  alpha = 0.5,
  point_size = 3,
  linewidth = 2
)
}
\arguments{
<<<<<<< HEAD
\item{p}{\strong{numeric factor} \cr
The values for the p-values.}

\item{FC}{\strong{numeric factor} \cr
The values for the fold changes.}

\item{significance_category}{\strong{character factor} \cr
The significance categories for the volcano plot.}

\item{log_base_fc}{\strong{numeric} \cr
The base for the fold changes log-transformation.}

\item{log_base_p}{\strong{numeric} \cr
The base for the p-values log-transformation.}

\item{thres_p}{\strong{numeric} \cr
The threshold for the p-values.}

\item{thres_fc}{\strong{numeric} \cr
The threshold for the fold changes.}

\item{colour1}{\strong{character} \cr
The color for not significant proteins.}

\item{colour2}{\strong{character} \cr
The color for significant proteins.}

\item{colour3}{\strong{character} \cr
The color for significant proteins after FDR correction.}

\item{symmetric_x}{\strong{logical} \cr
If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is defined).}

\item{legend_position}{\strong{character} \cr
The positioning of the legend.
Options are "none", "left", "right", "bottom", "top" and "inside".}

\item{base_size}{\strong{numeric} \cr
The base size of the font.}

\item{xlim}{\strong{numeric} \cr
The limits for x-axis.}

\item{ylim}{\strong{numeric} \cr
The limits for y-axis.}

\item{alpha}{\strong{numeric} \cr
The transparency of the data points.}

\item{point_size}{\strong{numeric} \cr
The size of the data points.}
}
\value{
A ggplot showing the volcano plot.
=======
\item{p}{vector of p-values (before p-value correction)}

\item{FC}{vector of fold changes}

\item{significance_category}{vector of significance categories ("not significant", "significant", "significant after FDR correction")}

\item{log_base_fc}{base for logarithm of the fold change (default is 2).}

\item{log_base_p}{base for logarithm of the p-value (default is 10, the negative log-p-value is plotted, e.g. -log10(p)).}

\item{thres_p}{threshold for p-value (default is 0.05)}

\item{thres_fc}{threshold for fold change (default is 2)}

\item{colour1}{colour for not significant points, default is grey}

\item{colour2}{colour for significant points, default is black}

\item{colour3}{colour for significant points after FDR correction, default is orange}

\item{symmetric_x}{if TRUE, x-axis will be made symmetric, default is FALSE}

\item{legend_position}{position of the legen, default is "bottom"}

\item{base_size}{base size for ggplot2 theme, default is NULL}

\item{xlim}{x-axis limits, default is NULL, i.e. will be determined automatically}

\item{ylim}{y-axis limits, default is NULL, i.e. will be determined automatically}
}
\value{
Volcano plot as ggplot object
>>>>>>> 87feb7e7ecaac8ef8743c891d467d40e48c27c9c
}
\description{
Simple volcano plot from p-value, fold change and significance category
}
<<<<<<< HEAD
\seealso{
\code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=VolcanoPlot_ANOVA]{VolcanoPlot_ANOVA()}}
=======
\examples{
# TODO
>>>>>>> 87feb7e7ecaac8ef8743c891d467d40e48c27c9c
}
