% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{VolcanoPlot_ANOVA}
\alias{VolcanoPlot_ANOVA}
\title{Volcano Plots for an ANOVA result.}
\usage{
VolcanoPlot_ANOVA(
  RES,
  columnname_p_ANOVA = "p.anova",
  columnname_p_ANOVA_adj = "p.anova.fdr",
  columns_FC,
  columns_p_posthoc,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  add_labels = FALSE
)
}
\arguments{
\item{RES}{\strong{data.frame} \cr
The results from an ANOVA.}

\item{columnname_p_ANOVA}{\strong{character} \cr
The column name for the p-values.}

\item{columnname_p_ANOVA_adj}{\strong{character} \cr
The column name for the adjusted p-values.}

\item{columns_FC}{\strong{integer vector} \cr
The column indices of the fold changes.}

\item{columns_p_posthoc}{\strong{integer vector} \cr
The column indices for the posthoc p-values.}

\item{log_base_fc}{\strong{numeric} \cr
The base for the fold changes log-transformation.}

\item{log_base_p}{\strong{numeric} \cr
The base for the p-values log-transformation.}

\item{thres_p}{\strong{numeric} \cr
The threshold for p-value.}

\item{thres_fc}{\strong{numeric} \cr
The threshold for fold change.}

\item{colour1}{\strong{character} \cr
The color for not significant proteins.}

\item{colour2}{\strong{character} \cr
The color for significant proteins.}

\item{colour3}{\strong{character} \cr
The color for significant proteins after FDR correction.}

\item{symmetric_x}{\strong{logical} \cr
If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is defined).}

\item{legend_position}{\strong{character} \cr
The positioning of the legend.
Options are "none", "left", "right", "bottom", "top" and "inside".}

\item{base_size}{\strong{numeric} \cr
The base size for theme.}

\item{xlim}{\strong{numeric} \cr
The limits for x-axis.}

\item{ylim}{\strong{numeric} \cr
The limits for y-axis.}

\item{add_labels}{\strong{logical} \cr
If \code{TRUE}, labels will be added.}
}
\value{
A list of ggplots of the volcano plot from an ANOVA result.
}
\description{
Volcano Plots for an ANOVA result.
}
\seealso{
\code{\link[=VolcanoPlot]{VolcanoPlot()}}, \code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=add_labels]{add_labels()}}
}
