% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{VolcanoPlot_ANOVA}
\alias{VolcanoPlot_ANOVA}
\title{Volcano Plots for an ANOVA result.}
\usage{
VolcanoPlot_ANOVA(
  RES,
  columnNamePAnova = "p.anova",
  columnNamePAnovaAdj = "p.anova.fdr",
  columnsFC,
  columnsPPosthoc,
  symmetricX = FALSE,
  baseSize = NULL,
  addLabels = FALSE,
  ...
)
}
\arguments{
\item{RES}{\strong{data.frame} \cr
The results from an ANOVA.}

\item{columnNamePAnova}{\strong{character(1)} \cr
The column name for the p-values.}

\item{columnNamePAnovaAdj}{\strong{character(1)} \cr
The column name for the adjusted p-values.}

\item{columnsFC}{\strong{integer} \cr
The column indices of the fold changes.}

\item{columnsPPosthoc}{\strong{integer} \cr
The column indices for the posthoc p-values.}

\item{symmetricX}{\strong{logical(1)} \cr
If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is
defined). Default is FALSE.}

\item{baseSize}{\strong{numeric(1)} \cr
The base size for the theme. Default is NULL.}

\item{addLabels}{\strong{logical(1)} \cr
If \code{TRUE}, labels will be added. Default is FALSE.}

\item{...}{Additional arguments passed on to \code{\link[=VolcanoPlot]{VolcanoPlot()}}, e.g.
\code{logBaseFC}, \code{logBaseP}, \code{thresP}, \code{thresFC},
\code{colour1}, \code{colour2}, \code{colour3}, \code{legendPosition},
\code{xlim}, \code{ylim}, \code{alpha}, \code{pointSize}, \code{linewidth}
or \code{showThresLine}.}
}
\value{
A list of ggplots of the volcano plot from an ANOVA result.
}
\description{
Volcano Plots for an ANOVA result.
}
\seealso{
\code{\link[=VolcanoPlot]{VolcanoPlot()}}, \code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=add_labels]{add_labels()}}
}
