% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot_ANOVA}
\alias{VolcanoPlot_ANOVA}
\title{Volcano Plots for ANOVA}
\usage{
VolcanoPlot_ANOVA(
  RES,
  columnname_p_ANOVA = "p.anova",
  columnname_p_ANOVA_adj = "p.anova.fdr",
  columns_FC,
  columns_p_posthoc,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  add_labels = FALSE
)
}
\arguments{
\item{RES}{result of ProtStatsWF::ANOVA() or other ANOVA function}

\item{columnname_p_ANOVA}{column name of the p-values from the ANOVA}

\item{columnname_p_ANOVA_adj}{column name of the adjusted p-values from the ANOVA}

\item{columns_FC}{vector of column numbers for the fold changes}

\item{columns_p_posthoc}{vector of column numbers for the posthoc p-values}

\item{log_base_fc}{base for log2 transformation of fold changes}

\item{log_base_p}{base for log10 transformation of p-values}

\item{thres_p}{threshold for p-values}

\item{thres_fc}{threshold for fold changes}

\item{colour1}{colour for not significant proteins}

\item{colour2}{colour for significant proteins}

\item{colour3}{colour for significant proteins after FDR correction}

\item{symmetric_x}{logical, if TRUE, x-axis will be made symmetric}

\item{legend_position}{position of the legend}

\item{base_size}{base size for the ggplot theme}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{add_labels}{add labels to interesting proteins in the volcano plot}
}
\value{
list of volcano plots as ggplot objects
}
\description{
Volcano Plots for ANOVA
}
\examples{
# TODO
}
