% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot_ANOVA}
\alias{VolcanoPlot_ANOVA}
\title{Volcano Plots for an ANOVA result.}
\usage{
VolcanoPlot_ANOVA(
  RES,
  columnname_p_ANOVA = "p.anova",
  columnname_p_ANOVA_adj = "p.anova.fdr",
  columns_FC,
  columns_p_posthoc,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  add_labels = FALSE
)
}
\arguments{
<<<<<<< HEAD
\item{RES}{\strong{data.frame} \cr
The result of ProtStatsWF::ANOVA() or other ANOVA function.}

\item{columnname_p_ANOVA}{\strong{character} \cr
The column name for the p-values of the ANOVA.}

\item{columnname_p_ANOVA_adj}{\strong{character} \cr
The column name for the adjusted p-values of the ANOVA.}

\item{columns_FC}{\strong{integer vector} \cr
The column indices of the fold changes.}

\item{columns_p_posthoc}{\strong{integer vector} \cr
The column indices for the posthoc p-values.}

\item{log_base_fc}{\strong{numeric} \cr
The base for the fold changes log-transformation.}

\item{log_base_p}{\strong{numeric} \cr
The base for the p-values log-transformation.}

\item{thres_p}{\strong{numeric} \cr
The threshold for p-value.}

\item{thres_fc}{\strong{numeric} \cr
The threshold for fold change.}

\item{colour1}{\strong{character} \cr
The color for not significant proteins.}

\item{colour2}{\strong{character} \cr
The color for significant proteins.}

\item{colour3}{\strong{character} \cr
The color for significant proteins after FDR correction.}

\item{symmetric_x}{\strong{logical} \cr
If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is defined).}

\item{legend_position}{\strong{character} \cr
The positioning of the legend.
Options are "none", "left", "right", "bottom", "top" and "inside".}

\item{base_size}{\strong{numeric} \cr
The base size for theme.}

\item{xlim}{\strong{numeric} \cr
The limits for x-axis.}

\item{ylim}{\strong{numeric} \cr
The limits for y-axis.}

\item{add_labels}{\strong{logical} \cr
If \code{TRUE}, labels will be added.}
}
\value{
A list of ggplots of the volcano plot from an ANOVA result.
=======
\item{RES}{result of ProtStatsWF::ANOVA() or other ANOVA function}

\item{columnname_p_ANOVA}{column name of the p-values from the ANOVA}

\item{columnname_p_ANOVA_adj}{column name of the adjusted p-values from the ANOVA}

\item{columns_FC}{vector of column numbers for the fold changes}

\item{columns_p_posthoc}{vector of column numbers for the posthoc p-values}

\item{log_base_fc}{base for log2 transformation of fold changes}

\item{log_base_p}{base for log10 transformation of p-values}

\item{thres_p}{threshold for p-values}

\item{thres_fc}{threshold for fold changes}

\item{colour1}{colour for not significant proteins}

\item{colour2}{colour for significant proteins}

\item{colour3}{colour for significant proteins after FDR correction}

\item{symmetric_x}{logical, if TRUE, x-axis will be made symmetric}

\item{legend_position}{position of the legend}

\item{base_size}{base size for the ggplot theme}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{add_labels}{add labels to interesting proteins in the volcano plot}
}
\value{
list of volcano plots as ggplot objects
>>>>>>> 87feb7e7ecaac8ef8743c891d467d40e48c27c9c
}
\description{
Volcano Plots for an ANOVA result.
}
\seealso{
\code{\link[=VolcanoPlot]{VolcanoPlot()}}, \code{\link[=VolcanoPlot_ttest]{VolcanoPlot_ttest()}}, \code{\link[=add_labels]{add_labels()}}
}
\examples{
# TODO
}
