% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot_ANOVA}
\alias{VolcanoPlot_ANOVA}
\title{Volcano Plots for ANOVA}
\usage{
VolcanoPlot_ANOVA(
  RES,
  columnname_p_ANOVA = "p.anova",
  columnname_p_ANOVA_adj = "p.anova.fdr",
  columns_FC,
  columns_p_posthoc,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  add_labels = FALSE
)
}
\arguments{
\item{ylim}{}
}
\description{
Volcano Plots for ANOVA
}
