% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot_ANOVA}
\alias{VolcanoPlot_ANOVA}
\title{Volcano Plots for ANOVA}
\usage{
VolcanoPlot_ANOVA(
  RES,
  columnname_p_ANOVA = "p.anova",
  columnname_p_ANOVA_adj = "p.anova.fdr",
  columns_FC,
  columns_p_posthoc,
  log_base_fc = 2,
  log_base_p = 10,
  thres_p = 0.05,
  thres_fc = 2,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  symmetric_x = FALSE,
  legend_position = "bottom",
  base_size = NULL,
  xlim = NULL,
  ylim = NULL,
  add_labels = FALSE
)
}
\arguments{
\item{RES}{Results table from ANOVA}

\item{columnname_p_ANOVA}{Column names for the p-values.}

\item{columnname_p_ANOVA_adj}{Column names for the adjusted p-values.}

\item{columns_FC}{Vector containing the indices of the fold change columns.}

\item{columns_p_posthoc}{Column names for the posthoc p-values.}

\item{log_base_fc}{Numeric of the base for the fold changes log-transformation}

\item{log_base_p}{Numeric of the base for the p-values log-transformation}

\item{thres_p}{Numeric of the threshold for the p-values}

\item{thres_fc}{Numeric of the threshold for the fold changes}

\item{colour1}{Character of the color for not significant proteins}

\item{colour2}{Character of the color for significant proteins}

\item{colour3}{Character of the color for significant proteins after FDR correction}

\item{symmetric_x}{If TRUE, x-axis limits will be made symmetric (not used if xlim is defined)}

\item{legend_position}{Character of the positioning of the legend, e.g. "bottom" or "right"}

\item{base_size}{Numeric containing the base size of the font}

\item{xlim}{The limits for x-axis. Default is NULL}

\item{ylim}{The limits for y-axis. Default is NULL}

\item{add_labels}{If TRUE, labels are added to the plot.}
}
\value{
A list of ggplots
}
\description{
Volcano Plots for ANOVA
}
