% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{VolcanoPlot_ttest}
\alias{VolcanoPlot_ttest}
\title{Volcano plot for a t-test result.}
\usage{
VolcanoPlot_ttest(
  RES,
  columnNameP = "p",
  columnNamePadj = "padj",
  columnNameFC = "FC",
  thresFC = 2,
  thresP = 0.05,
  logBaseFC = 2,
  logBaseP = 10,
  isFCLog = FALSE,
  isPLog = FALSE,
  showThresLine = TRUE,
  groupName1 = "group1",
  groupName2 = "group2",
  plotHeight = 15,
  plotWidth = 15,
  plotDPI = 300,
  plotDevice = "pdf",
  outputPath = NULL,
  suffix = NULL,
  addAnnotation = TRUE,
  ...
)
}
\arguments{
\item{RES}{\strong{data.frame} \cr
The results from a t-test.}

\item{columnNameP}{\strong{character} \cr
The column name for p-value.}

\item{columnNamePadj}{\strong{character} \cr
The columns name for adjusted p-value.}

\item{columnNameFC}{\strong{character} \cr
The column name for fold change.}

\item{thresFC}{\strong{numeric} \cr
The threshold for fold change.}

\item{thresP}{\strong{numeric} \cr
The threshold for p-value.}

\item{logBaseFC}{\strong{numeric} \cr
The base for the fold changes log-transformation.}

\item{logBaseP}{\strong{numeric} \cr
The base for the p-values log-transformation.}

\item{isFCLog}{\strong{logical} \cr
If \code{TRUE}, fold change is already log-transformed.}

\item{isPLog}{\strong{logical} \cr
If \code{TRUE}, p-value is already log-transformed.}

\item{showThresLine}{\strong{logical} \cr
If \code{TRUE}, threshold lines will be shown.}

\item{groupName1}{\strong{character} \cr
The name of first group.}

\item{groupName2}{\strong{character} \cr
The name of second group.}

\item{plotHeight}{\strong{numeric} \cr
The height of plot.}

\item{plotWidth}{\strong{numeric} \cr
The width of plot.}

\item{plotDPI}{\strong{integer} \cr
The resolution of plot.}

\item{plotDevice}{\strong{character} \cr
The plot device that is used for the resulting plot.
Options are "pdf" and "png".}

\item{outputPath}{\strong{character} \cr
The path for output file.}

\item{suffix}{\strong{character} \cr
The suffix for output file.}

\item{addAnnotation}{\strong{logical} \cr
If \code{TRUE}, annotation will be added.}

\item{...}{Additional arguments for the plot.}
}
\value{
A ggplot object of the volcano plot from a ttest result.
}
\description{
Volcano plot for a t-test result.
}
\seealso{
\code{\link[=VolcanoPlot]{VolcanoPlot()}}, \code{\link[=VolcanoPlot_ANOVA]{VolcanoPlot_ANOVA()}}, \code{\link[=add_labels]{add_labels()}}
}
