% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlots.R
\name{VolcanoPlot_ttest}
\alias{VolcanoPlot_ttest}
\title{Volcano plot for a t-test result}
\usage{
VolcanoPlot_ttest(
  RES,
  columnname_p = "p",
  columnname_padj = "padj",
  columnname_FC = "FC",
  log_base_fc = 2,
  log_base_p = 10,
  is_FC_log = FALSE,
  is_p_log = FALSE,
  thres_fc = 2,
  thres_p = 0.05,
  show_thres_line = TRUE,
  colour1 = "grey",
  colour2 = "black",
  colour3 = "orange",
  groupname1 = "group1",
  groupname2 = "group2",
  xlim = NULL,
  ylim = NULL,
  symmetric_x = FALSE,
  legend_position = "bottom",
  plot_height = 15,
  plot_width = 15,
  plot_dpi = 300,
  plot_device = "pdf",
  output_path = NULL,
  suffix = NULL,
  base_size = NULL,
  add_annotation = TRUE
)
}
\arguments{
\item{RES}{A data.frame of the result table from t-test}

\item{columnname_p}{The column name for p-value}

\item{columnname_padj}{The columns name for adjusted p-value}

\item{columnname_FC}{The column name for fold change}

\item{log_base_fc}{The base for log transformation of fold change}

\item{log_base_p}{The base for log transformation of p-value}

\item{is_FC_log}{If \code{TRUE}, fold change is already log-transformed}

\item{is_p_log}{If \code{TRUE}, p-value is already log-transformed}

\item{thres_fc}{The threshold for fold change}

\item{thres_p}{The threshold for p-value}

\item{show_thres_line}{If \code{TRUE}, threshold lines will be shown}

\item{colour1}{The colour for not significant proteins}

\item{colour2}{The colour for significant proteins}

\item{colour3}{The colour for significant proteins after FDR correction}

\item{groupname1}{The name of first group}

\item{groupname2}{The name of second group}

\item{xlim}{The limits for x-axis}

\item{ylim}{The limits for y-axis}

\item{symmetric_x}{If \code{TRUE}, x-axis limits will be made symmetric (not used if xlim is defined)}

\item{legend_position}{The position of the legend (e.g., "bottom" or "right")}

\item{plot_height}{The height of plot}

\item{plot_width}{The width of plot}

\item{plot_dpi}{The resolution of plot}

\item{plot_device}{The plot device that is used for the resulting plot.}

\item{output_path}{The path for output file}

\item{suffix}{The suffix for output file}

\item{base_size}{The base size for theme}

\item{add_annotation}{If \code{TRUE}, annotation should be added}
}
\value{
ggplot object
}
\description{
Volcano plot for a t-test result
}
\examples{
# TODO
}
