% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_clustering.R
\name{workflow_clustering}
\alias{workflow_clustering}
\title{Workflow that clusters proteins for similar patterns over the samples and
produces a heatmap and lineplots}
\usage{
workflow_clustering(
  data_path,
  output_path,
  intensity_columns,
  nr_clusters = NULL,
  cluster_colours = NULL,
  row_split = TRUE,
  dist_method = "correlation",
  colour_dend = TRUE,
  suffix = "",
  plot_height_heatmap = 15,
  plot_width_heatmap = 15,
  plot_height_lineplot = 20,
  plot_width_lineplot = 25,
  plot_dpi = 300,
  column_name_protein = "Protein",
  ...
)
}
\arguments{
\item{data_path}{\strong{character} \cr
The path to an .xlsx file containing the input data.}

\item{output_path}{\strong{character} \cr
The path to the output folder.}

\item{intensity_columns}{\strong{integer vector} \cr
The numbers of the intensity columns in the table.}

\item{nr_clusters}{\strong{integer(1)} \cr
Number of clusters to cut the dendrogram into. If \code{NULL} (default),
the optimal number of clusters is determined based on silhouette values
using the \code{\link[dendextend]{find_k}} function.}

\item{cluster_colours}{\strong{character vector} \cr
Colours to use for the different clusters. If \code{NULL} (default),
the default ggplot color palette is used.}

\item{row_split}{\strong{logical(1)} \cr If TRUE, there will be space between row clusters in the heatmap.}

\item{dist_method}{\strong{character(1)} \cr
Distance measure to use for the hierarchical clustering. In principle,
all methods available in \code{\link[amap]{Dist}} are possible, however
correlation-based metrics like "correlation",  "pearson" or "spearman"
are recommended. The default is "correlation", which uses the centere
Pearson correlation.}

\item{colour_dend}{\strong{logical(1)} \cr
If \code{TRUE} (default), the branches of the dendrogram are coloured
according to the clusters, using the defined cluster_colours.}

\item{suffix}{\strong{character} \cr
The suffix for the output files. It needs to start with an underscore.}

\item{plot_height_heatmap}{\strong{numeric} \cr
The height for the heatmap in cm.}

\item{plot_width_heatmap}{\strong{numeric} \cr
The width for the heatmap in cm.}

\item{plot_height_lineplot}{\strong{numeric} \cr
The height for the lineplots in cm.}

\item{plot_width_lineplot}{\strong{numeric} \cr
The width for the lineplots in cm.}

\item{plot_dpi}{\strong{numeric} \cr
The plot resolution for the heatmap.}

\item{column_name_protein}{\strong{character(1)} \cr
The name of the column containing the protein identifiers.}

\item{...}{Additional parameters passed to \code{\link[ProtStatsWF]{heatmap}}.}
}
\value{
Nothing, but saves a heatmap, a set of lineplots (one per cluster)
and a cluster table to the output folder.
}
\description{
Workflow that clusters proteins for similar patterns over the samples and
produces a heatmap and lineplots
}
