% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_QC.R
\name{workflow_QC}
\alias{workflow_QC}
\title{QC Workflow (QCQuant)}
\usage{
workflow_QC(
  D,
  groupColumn = NULL,
  group2Column = NULL,
  groupColours = NULL,
  outPath,
  outType = "xlsx",
  suffix = "",
  NAOut = "NA",
  verbose = TRUE,
  baseSize = 15,
  plotDevice = "pdf",
  plotHeight_BP_VV = 10,
  plotWidth_BP_VV = 15,
  plotHeight_PCA_MA = 15,
  plotWidth_PCA_MA = 15,
  plotDPI = 300,
  boxplotMethod = "boxplot",
  MAMaxPlots = 5000,
  MAAlpha = 1,
  PCAImputeMethod = "mean",
  PCAPropNA = 0,
  PCAScale = TRUE,
  PCAAlpha = 1,
  PCALabel = FALSE,
  PCALabelSeed = NA,
  PCALabelSize = 4,
  PCAXlim = NULL,
  PCAYlim = NULL,
  PCAPointSize = 4
)
}
\arguments{
\item{D}{\strong{list} Result from \code{\link[=prepareDatSE]{prepareDatSE()}} containing the imported data.}

\item{groupColumn}{\strong{character(1)} Column name in the colData(D$SE) that contains the group
information for colouring the plots. Default is NULL (no group information for colouring).}

\item{group2Column}{\strong{character(1)} Column name in the colData(D$SE) that
contains a secondary group information (used in PCA_plot as shape). Default is NULL.}

\item{groupColours}{\strong{character} Vector of colours for the groups. Default is NULL
(default ggplot2 colour palette will be used.)}

\item{outPath}{\strong{character(1)} Path to save the output files. The corresponding folder must exist.}

\item{outType}{\strong{character(1)} Type of output file. Default is xlsx.}

\item{suffix}{\strong{character(1)} Suffix to add to the output file names. Default is "".}

\item{NAOut}{\strong{character(1)} String to represent missing values in the output files. Default is "NA".}

\item{verbose}{\strong{logical(1)} Whether to print messages during the workflow. Default is TRUE.}

\item{baseSize}{\strong{numeric(1)} Base size for the plots. Default is 15.}

\item{plotDevice}{\strong{character(1)} Device to use for saving plots. Default is "pdf".}

\item{plotHeight_BP_VV}{\strong{numeric(1)} Height of the boxplot and valid value plots in cm. Default is 10.}

\item{plotWidth_BP_VV}{\strong{numeric(1)} Width of the boxplot and valid value plots in cm. Default is 15.}

\item{plotHeight_PCA_MA}{\strong{numeric(1)} Height of the PCA and MA plots in cm. Default is 15.}

\item{plotWidth_PCA_MA}{\strong{numeric(1)} Width of the PCA and MA plots in cm. Default is 15.}

\item{plotDPI}{\strong{numeric(1)} DPI for the plots. Default is 300.}

\item{boxplotMethod}{\strong{character(1)} Method for boxplot. Default is "boxplot".}

\item{MAMaxPlots}{\strong{numeric(1)} Maximum number of MA plots to be drawn. Default is 5000.}

\item{MAAlpha}{\strong{numeric(1)} Alpha value for MA plots. Default is 1.}

\item{PCAImputeMethod}{\strong{character(1)} Method for missing value imputation for the PCA. Default is "mean".}

\item{PCAPropNA}{\strong{numeric(1)} Allowed proportion of missing values for PCA. Default is 0.
Proteins with more missing values will be removed from the PCA. The others will be imputed.}

\item{PCAScale}{\strong{logical(1)} Whether to scale data for the PCA. Default is TRUE.}

\item{PCAAlpha}{\strong{numeric(1)} Alpha value (transparency) for points in the PCA. Default is 1.}

\item{PCALabel}{\strong{logical(1)} Whether to label PCA points. Default is FALSE.}

\item{PCALabelSeed}{\strong{numeric(1)} Seed for data point labels in the PCA. Default is NA.}

\item{PCALabelSize}{\strong{numeric(1)} Size of data point labels in the PCA. Default is 4.}

\item{PCAXlim}{\strong{numeric(2)} Limits for the x-axis in the PCA plot. Default is NULL.}

\item{PCAYlim}{\strong{numeric(2)} Limits for the y-axis in the PCA plot. Default is NULL.}

\item{PCAPointSize}{\strong{numeric(1)} Size of data points in the PCA. Default is 4.}
}
\value{
The workflow saves several plots and files and returns a message log of the workflow.
}
\description{
Workflow for quality control of quantitative proteomics data
}
\details{
This function performs quality control of quantitative proteomics data.
The following plots are generated: a valid value plot, boxplots, MA-plots and a PCA plot.
}
\seealso{
Functions used in this workflow:
\code{\link[=prepareDataSE]{prepareDataSE()}}, \code{\link[=ValidValuePlot]{ValidValuePlot()}}, \code{\link[=Boxplots]{Boxplots()}}, \code{\link[=MA_Plots]{MA_Plots()}}, \code{\link[=PCA_Plot]{PCA_Plot()}}.
}
