% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_QC.R
\name{workflow_QC}
\alias{workflow_QC}
\title{The main workflow for quality control of quantitative proteomics data}
\usage{
workflow_QC(
  data_path,
  output_path,
  intensity_columns,
  normalization_method = "loess",
  lts_quantile = 0.8,
  use_groups = TRUE,
  na_strings = c("NA", "NaN", "Filtered", "#NV"),
  zero_to_NA = TRUE,
  do_log_transformation = TRUE,
  log_base = 2,
  groupvar_name = "Group",
  group_colours = NULL,
  base_size = 15,
  plot_device = "pdf",
  plot_height = 10,
  plot_width = 15,
  plot_dpi = 300,
  suffix = "_",
  boxplot_method = "boxplot",
  generate_MAplots = TRUE,
  MA_maxPlots = 5000,
  MA_alpha = FALSE,
  MA_sampling = 1,
  PCA_impute = FALSE,
  PCA_impute_method = "mean",
  PCA_propNA = 0,
  PCA_scale. = TRUE,
  PCA_PCx = 1,
  PCA_PCy = 2,
  PCA_groupvar1_name = "group",
  PCA_alpha = 1,
  PCA_label = FALSE,
  PCA_label_seed = NA,
  PCA_label_size = 4,
  PCA_xlim = NULL,
  PCA_ylim = NULL,
  PCA_point.size = 4
)
}
\arguments{
\item{data_path}{A character containing the path to an .xlsx file.}

\item{output_path}{A character containing the path to an output folder.}

\item{intensity_columns}{An integer vector containing the intensity columns of the table.}

\item{normalization_method}{A character containing the method of normalization. The possible methods are no normalization "nonorm" or "median", "loess", "quantile" or "lts" normalization.}

\item{lts_quantile}{A numeric containing the quantile for the lts normalization if \code{normalization = "lts"}, default is 0.8.}

\item{use_groups}{If \code{TRUE}, group information encoded in the column names are used. Default is \code{TRUE}.}

\item{na_strings}{A character vector containing symbols to be recognized as missing values (with the exception of 0).}

\item{zero_to_NA}{If \code{TRUE}, 0 will be treated as missing value.}

\item{do_log_transformation}{If \code{TRUE}, the data will be log-transformed.}

\item{log_base}{A numeric containing the base used, if data is log-transformed.}

\item{groupvar_name}{A character containing the name for the group variable.}

\item{group_colours}{A character vector of hex codes for the group colors, if the data has groups. If \code{NULL}, a default color scale will be used.}

\item{base_size}{A numeric containing the base size of the font.}

\item{plot_device}{A character containing the type of the output file, e.g. "pdf" or "png".}

\item{plot_height}{A numeric of the plot height in cm.}

\item{plot_width}{A numeric of the plot width in cm.}

\item{plot_dpi}{A numeric of the "dots per inch" of the plot aka. the plot resolution.}

\item{suffix}{A character containing the suffix for the output files. Needs to start with an underscore.}

\item{boxplot_method}{A character containing the method used. Possible are "boxplot" and "violinplot". Default is "boxplot".}

\item{generate_MAplots}{If \code{TRUE}, MA plots will be generated, if \code{FALSE} they will not be generated (mostly for debugging purposes).}

\item{MA_maxPlots}{A numeric containing the maximum number of MA plots that should be generated. Defaults is 5000.}

\item{MA_alpha}{If \code{TRUE}, the data points of the MA plots will be transparent.}

\item{MA_sampling}{A numeric containing the sampling rate for MA-Plots. Useful to sample part of the data set for data sets on peptide/feature level with many data points.}

\item{PCA_impute}{If \code{TRUE}, missing values will be imputed.}

\item{PCA_impute_method}{A character containing the imputation method ("mean" or "median")}

\item{PCA_propNA}{A numeric of the proportion of allowed missing NAs for a protein, before it is discarded.}

\item{PCA_scale.}{If \code{TRUE}, the data will be scaled before computing the PCA.}

\item{PCA_PCx}{The principle component for the x-axis (default: 1).}

\item{PCA_PCy}{The principle component for the y-axis (default: 2).}

\item{PCA_groupvar1_name}{Titles of legends for colour.}

\item{PCA_alpha}{If \code{TRUE}, the data points of the PCA plot will be transparent.}

\item{PCA_label}{If \code{TRUE}, the samples will be labeled.}

\item{PCA_label_seed}{A numeric, which sets the seed for the label.}

\item{PCA_label_size}{A numeric containing the size of the sample labels.}

\item{PCA_xlim}{Limit of the x-axis.}

\item{PCA_ylim}{Limit of the y-axis.}

\item{PCA_point.size}{The size of the data points.}
}
\value{
saves several plots and excel files
}
\description{
The main workflow for quality control of quantitative proteomics data
}
\examples{
\dontrun{
in_path <- "/Users/thisuser/Documents/dataFolder/data.xlsx"
intensity_cols <- 3:17
out_path <- "/Users/thisuser/Documents/resultsFolder/"

result <- workflow_QC(data_path = in_path,
                      intensity_columns = intensity_cols,
                      output_path = out_path)
}

}
