% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_ttest.R
\name{workflow_ttest}
\alias{workflow_ttest}
\title{The workflow for t-test, ANOVA etc. of quantitative proteomics data}
\usage{
workflow_ttest(
  data_path,
  output_path,
  sample = NULL,
  paired = FALSE,
  var.equal = FALSE,
  log_before_test = TRUE,
  delog_for_FC = TRUE
)
}
\arguments{
\item{data_path}{A character containing the path to an .xlsx file.}

\item{output_path}{A character containing the path to an output folder.}

\item{sample}{A factor of  which column belongs to which sample.}

\item{paired}{If \code{TRUE}, a paired test will be done, otherwise an unpaired test.}

\item{var.equal}{If \code{TRUE}, the variances are assumed to be equal.}

\item{log_before_test}{If \code{TRUE}, the data will be log-transformed.}

\item{delog_for_FC}{If \code{TRUE}, the fold change will be calculated without the log-transformation.}
}
\value{
Message log of the workflow
}
\description{
The workflow for t-test, ANOVA etc. of quantitative proteomics data
}
\examples{
\dontrun{
in_path <- "/Users/thisuser/Documents/dataFolder/data.xlsx"
out_path <- "/Users/thisuser/Documents/resultsFolder/"

result <- workflow_ttest(data_path = in_path,
                      output_path = out_path)
}
}
