% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_ttest.R
\name{workflow_ttest}
\alias{workflow_ttest}
\title{The main workflow for quality control of quantitative proteomics data}
\usage{
workflow_ttest(data_path, output_path)
}
\arguments{
\item{data_path}{A character containing the path to an .xlsx file.}

\item{output_path}{A character containing the path to an output folder.}
}
\value{
TODO
}
\description{
The main workflow for quality control of quantitative proteomics data
}
\examples{
\dontrun{
in_path <- "/Users/thisuser/Documents/dataFolder/data.xlsx"
out_path <- "/Users/thisuser/Documents/resultsFolder/"

result <- workflow_ttest(data_path = in_path,
                      output_path = out_path)
}
}
