% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_ttest.R
\name{workflow_ttest}
\alias{workflow_ttest}
\title{The workflow for t-test of quantitative proteomics data}
\usage{
workflow_ttest(
  data_path,
  output_path,
  intensity_columns,
  paired = FALSE,
  var.equal = FALSE,
  log_before_test = TRUE,
  delog_for_FC = TRUE,
  significant_after_FDR = TRUE,
  max_valid_values_off = 0,
  min_valid_values_on = NULL,
  suffix = "",
  plot_device = "pdf",
  plot_height = 15,
  plot_width = 15,
  plot_dpi = 300
)
}
\arguments{
\item{data_path}{A character containing the path to an .xlsx file.}

\item{output_path}{A character containing the path to an output folder.}

\item{intensity_columns}{An integer vector containing the intensity columns of the table.}

\item{paired}{If \code{TRUE}, a paired test will be done, otherwise an unpaired test.}

\item{var.equal}{If \code{TRUE}, the variances are assumed to be equal.}

\item{log_before_test}{If \code{TRUE}, the data will be log-transformed.}

\item{delog_for_FC}{If \code{TRUE}, the fold change will be calculated without the log-transformation.}

\item{significant_after_FDR}{If \code{TRUE}, candidates for the boxplots and heatmap need to be significant after FDR correction, otherwise all significant candidates will be used.}

\item{max_valid_values_off}{A numeric of the maximum number of valid values to be an off protein}

\item{min_valid_values_on}{A numeric of the minimum number of valid values to be an on protein}

\item{suffix}{A character if the file names should contain a suffix.}

\item{plot_device}{A character containing the type of the output file, e.g. "pdf" or "png".}

\item{plot_height}{A numeric of the plot height in cm.}

\item{plot_width}{A numeric of the plot width in cm.}

\item{plot_dpi}{A numeric of the "dots per inch" of the plot aka. the plot resolution.}
}
\value{
Message log of the workflow
}
\description{
The workflow for t-test of quantitative proteomics data
}
\examples{
\dontrun{
result <- workflow_ttest(data_path = "/Users/thisuser/Documents/dataFolder/data.xlsx",
                         output_path = "/Users/thisuser/Documents/resultsFolder/",
                         intensity_columns = 3:8)
}
}
